### 2024-07-03 #6960 @editorialbot
[PRE REVIEW]: OpenMD: A parallel molecular dynamics engine for complex systems and interfaces
**Submitting author:** @gezelter (<a href="http://orcid.org/0000-0002-2935-3163">J. Daniel Gezelter</a>)
**Repository:** https://github.com/OpenMD/OpenMD/
**Branch with paper.md** (empty if default branch):
**Version:** v3.1
