### 2024-07-08 #6960 @Kevin-Mattheus-Moerman
@gezelter, thanks for this submission. I am the AEiC on this track and here to help process the initial steps. Before we proceed, please can you have a look at the following points:

- [ ] Please study the above reference check :point_up: and see if you can address any of the reported potential DOI issues. You can add/amend DOI entries in your `.bib` file, and call `@editorialbot check references` here to check them again. If the references truly do not have DOI's you may leave them.
