### 2024-07-08 #6960 @crdrisko
Hello @Kevin-Mattheus-Moerman, we have verified that the three articles mentioned *do not* have DOI's associated with them. In lieu of this, we've included urls for these articles, but we can remove them if it goes against the journal's formatting. Also, our `CONTRIBUTING.md` is located in the .github directory, [here](https://github.com/OpenMD/OpenMD/blob/main/.github/CONTRIBUTING.md).
