### 2024-07-08 #6960 @gezelter
@crdrisko mentions the DOI issue - the three papers are from journals that ceased publishing prior to the DOI system. Two of the papers are in the Computational Collaboration Project No 5 (CCP5) information newsletter which does have a [digital archive](https://www.ccp5.ac.uk/newsletter/) and the Delaunay paper was influential enough that it has been [archived at MathNet](https://www.mathnet.ru/eng/im4937).
