### 2024-07-08 ea24e6f
Update README.md with JOSS status
