### 2024-07-09 2c2ce5b
Fixing undeclared variable that were causing issues in time correlation functions
