### 2024-07-11 6ea57bd
Using a better naming convention for variables in SPF
