### 2024-07-12 250a8c0
Applying a current density instead of particle flux for charged spf
