+++ src/rnemd/RNEMD.hpp
@@ -81,26 +81,27 @@ namespace OpenMD::RNEMD {
     enum RNEMDPrivilegedAxis { rnemdX = 0, rnemdY = 1, rnemdZ = 2 };
 
     enum RNEMDFluxType {
-      rnemdKE,          // translational kinetic energy flux
-      rnemdRotKE,       // rotational kinetic energy flux
-      rnemdFullKE,      // full kinetic energy flux
-      rnemdPx,          // flux of momentum along x axis
-      rnemdPy,          // flux of momentum along y axis
-      rnemdPz,          // flux of momentum along z axis
-      rnemdPvector,     // flux of momentum vector
-      rnemdLx,          // flux of angular momentum along x axis
-      rnemdLy,          // flux of angular momentum along y axis
-      rnemdLz,          // flux of angular momentum along z axis
-      rnemdLvector,     // flux of angular momentum vector
-      rnemdParticle,    // flux of particles
-      rnemdParticleKE,  // flux of particles and total kinetic energy
-      rnemdKePx,        // flux of translational KE and x-momentum
-      rnemdKePy,        // flux of translational KE and y-momentum
-      rnemdKePvector,   // full combo flying platter
-      rnemdKeLx,        // flux of translational KE and x-angular momentum
-      rnemdKeLy,        // flux of translational KE and y-angular momentum
-      rnemdKeLz,        // flux of translational KE and z-angular momentum
-      rnemdKeLvector,   // full combo spinning platter
+      rnemdKE,              // translational kinetic energy flux
+      rnemdRotKE,           // rotational kinetic energy flux
+      rnemdFullKE,          // full kinetic energy flux
+      rnemdPx,              // flux of momentum along x axis
+      rnemdPy,              // flux of momentum along y axis
+      rnemdPz,              // flux of momentum along z axis
+      rnemdPvector,         // flux of momentum vector
+      rnemdLx,              // flux of angular momentum along x axis
+      rnemdLy,              // flux of angular momentum along y axis
+      rnemdLz,              // flux of angular momentum along z axis
+      rnemdLvector,         // flux of angular momentum vector
+      rnemdParticle,        // flux of particles
+      rnemdParticleKE,      // flux of particles and total kinetic energy
+      rnemdCurrentDensity,  // flux of electrons/charge
+      rnemdKePx,            // flux of translational KE and x-momentum
+      rnemdKePy,            // flux of translational KE and y-momentum
+      rnemdKePvector,       // full combo flying platter
+      rnemdKeLx,            // flux of translational KE and x-angular momentum
+      rnemdKeLy,            // flux of translational KE and y-angular momentum
+      rnemdKeLz,            // flux of translational KE and z-angular momentum
+      rnemdKeLvector,       // full combo spinning platter
       rnemdUnknownFluxType
     };
 
