### 2024-07-12 f6ad450
Changes for cmake 3.30 with FetchContent_Populate.  Formatting update to CMakeLists.txt
