+++ CMakeLists.txt
@@ -1,194 +1,218 @@
 # Top level cmake script for OpenMD.
 cmake_minimum_required(VERSION 3.20)
 
-project(OpenMD VERSION 3.1.0 LANGUAGES CXX)
+project(
+  OpenMD
+  VERSION 3.1.0
+  LANGUAGES CXX)
 
 set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
-if (CMAKE_CXX_COMPILER_ID STREQUAL "IntelLLVM")
-  SET(CMAKE_CXX_FLAGS_RELEASE "-O3 -ffp-model=precise")
-  SET(CMAKE_CXX_FLAGS_DEBUG "-g -ffp-model=strict")
+if(CMAKE_CXX_COMPILER_ID STREQUAL "IntelLLVM")
+  set(CMAKE_CXX_FLAGS_RELEASE "-O3 -ffp-model=precise")
+  set(CMAKE_CXX_FLAGS_DEBUG "-g -ffp-model=strict")
 endif()
 
 list(APPEND CMAKE_MODULE_PATH "${CMAKE_CURRENT_SOURCE_DIR}/cmake/modules")
 
-if (NOT CMAKE_BUILD_TYPE)
+if(NOT CMAKE_BUILD_TYPE)
   set(CMAKE_BUILD_TYPE "Release")
 endif()
 
-IF(WIN32)
-   ADD_DEFINITIONS(/D_CRT_SECURE_NO_WARNINGS)
-ENDIF(WIN32)
+if(WIN32)
+  add_definitions(/D_CRT_SECURE_NO_WARNINGS)
+endif(WIN32)
 
 include(FetchContent)
 
 find_package(RapidJSON)
 
-if (RAPIDJSON_FOUND)
-   include_directories(${RAPIDJSON_INCLUDE_DIRS})
-   message(STATUS "Adding include for json to ${RAPIDJSON_INCLUDE_DIRS}")
-   set(RAPID_JSON 1)
+if(RAPIDJSON_FOUND)
+  include_directories(${RAPIDJSON_INCLUDE_DIRS})
+  message(STATUS "Adding include for json to ${RAPIDJSON_INCLUDE_DIRS}")
+  set(RAPID_JSON 1)
 else()
   find_package(nlohmann_json)
-  
-  if (nlohmann_json_FOUND)
+
+  if(nlohmann_json_FOUND)
     set(NLOHMANN_JSON 1)
   else()
-   	message(STATUS "Fetching RapidJSON")
-	  FetchContent_Declare(rapidjson
-		                     GIT_REPOSITORY https://github.com/Tencent/rapidjson.git
-		                     GIT_TAG        master)
-	  FetchContent_GetProperties(rapidjson)
-	
-    if (NOT rapidjson_POPULATED)
-	    FetchContent_Populate(rapidjson)
-	  endif()
-
-	  include_directories(${rapidjson_SOURCE_DIR}/include)
-	  message(STATUS "Adding include for json to ${rapidjson_SOURCE_DIR}/include")
-	  set(RAPID_JSON 1)
+    set(RAPIDJSON_BUILD_DOC
+        OFF
+        CACHE BOOL "")
+    set(RAPIDJSON_BUILD_EXAMPLES
+        OFF
+        CACHE BOOL "")
+    set(RAPIDJSON_BUILD_TESTS
+        OFF
+        CACHE BOOL "")
+    set(RAPIDJSON_BUILD_THIRDPARTY_GTEST
+        OFF
+        CACHE BOOL "")
+    message(STATUS "Fetching RapidJSON")
+    FetchContent_Declare(
+      rapidjson
[…]
