### 2024-07-12 ba5dca1
Fixing missing checks in SPF.cpp for new parameters
