### 2024-07-16 8783a09
Updates to quiet down compilation warnings on Windows
