### 2024-07-18 a14e774
Reducing warnings on VisualStudio
