### 2024-07-18 a233eba
Function can be const
