### 2024-07-19 afffb4e
Updating build with qhull on Visual Studio 2022, initializing local variables
