### 2024-07-20 8cef91b
Allowing SPF to proceed with an invalid state by selecting another particle
