### 2024-07-31 a5d9817
Converting waterBoxer from Perl to Python
