### 2024-08-01 e0696c0
Converting waterSphere from Perl to Python, updating parts of waterBoxer to match
