### 2024-08-01 #7004 @samwaseda
Since the package contains quite a large number of files and lines, I think it's going to take a few days to check the functionality claimed in the paper.

In the meantime, I have two questions:

- Is there an example for the very first steps? I took a look at the directories inside `samples`, but the content doesn't follow a clear pattern, and the pdf manual gives a detailed instruction for the input files, but I couldn't figure out how to run it.
- Is it possible to locally install it on a computer without a sudo access? I managed to install it on my personal computer following the instructions on the homepage, but I would love to be able to run it on a different system. I'm sorry for probably asking a trivial question.
