### 2024-08-01 a5edd05
Fixing an issue in SPF where processors could get out of sync
