### 2024-08-01 efba5a7
Renaming and using markdown syntax for INSTALL
