+++ docs/INSTALL.md
@@ -1,0 +1,212 @@
+# Compiling OpenMD
+
+OpenMD is written in C++. Compiling is the process of turning this
+C++ into instructions that the computer’s processor can understand.
+
+## Requirements
+
+To build OpenMD, you need the following:
+
+* The source code for the latest release of OpenMD
+* A C++ compiler
+* CMake 3.15 or newer
+
+OpenMD uses CMake as its build system. CMake is an open source
+cross-platform build system from KitWare.
+
+You need to install CMake 3.15 or newer. This is available as a
+binary package from the KitWare website; alternatively, it may be
+available through your package manager (on Linux). If necessary, you
+can also compile it yourself from the source code.
+
+The following are optional when compiling OpenMD, but if they are not
+available some features will be missing:
+
+* OpenMPI – A very good implementation of the MPI-2 specification
+for parallel computing.  A version of the MPI library is required
+if you want to run the multi-processor version of OpenMD
+
+* perl and python - interpreted scripting languages that some of
+the OpenMD utilities use to parse and process data files. The python
+scripts now depend on NumPy and SciPy.
+
+* qhull – A computational geometry toolbox for computing convex
+hulls and Delaunay triangulations.  qhull is required for the
+LangevinHull integrator and for any of the tools that compute the
+Hull atoms or hull volumes of nanoparticles and clusters.
+
+* openbabel – a chemical toolbox for converting between different
+data formats.  This is required for building the atom2md program
+which helps prepare initial "metadata" or md files for
+simulations.
+
+* fftw - a library for computing discrete Fourier transforms.  This
+is required for surface undulation spectra (Hxy in
+staticProps). Get version 3.
+
+* zlib - required to support reading gzipped trajectory files
+
+You’ll also likely want to download and compile the following useful
+tools for interacting with the data:
+
+  * Jmol
+  * xmgr
+  * grace
+  * NumPy
+  * SciPy
+  * vmd
+
+If you are going to be extending or developing OpenMD, you’ll need
+the following tools:
+
+* antlr – our tool for parsing meta-data files.  You’ll want
+  version 2, not 3.  
+
+* gengetopt - a tool to generate C code to parse the command line
+  arguments argc and argv that are part of every C or C++ program
+
+## Basic build procedure
+
+The recommended way to build OpenMD is to use a separate source and
+build directory; for example, openmd-3.0 and build. The first step
+is to create these directories:
+
+```bash
+$ tar zxf openmd-3.0.tar.gz   # (this creates openmd-3.0)
+$ mkdir build
+```
+
+Now you need to run cmake to configure the build. The following will
+configure the build to use all of the default options:
+
+```bash
+$ cd build
+$ cmake ../openmd-3.0
+```
+
+If you need to specify a particular compiler, you can do that with
+environment variables before the cmake line
+
+```bash
+$ export CXX=/opt/local/lib/openmpi/bin/mpic++
+$ cmake ../openmd-3.0
+```
+
+If you need to specify an option, use the -D switch to cmake. For
+example, the following line sets the value of `CMAKE_INSTALL_PREFIX`
+and `CMAKE_BUILD_TYPE`:
+
+```bash
+$ cmake ../openmd-3.0 -DCMAKE_INSTALL_PREFIX=~/Tools -DCMAKE_BUILD_TYPE=DEBUG
+```
+
+We will discuss various possible options later.
+
+At this point, it would be a good idea to compile OpenMD:
+
+```bash
+$ make
+```
+
+Have a coffee while the magic happens. If you have a multi-processor
+machine and would prefer an espresso, try a parallel build instead:
+
+```bash
+$ make -j 4  
+```
+
+And finally, as root (or using sudo) you should install it:
+
+```bash
[…]
