### 2024-08-01 4971e93
Converted solvator from Perl to Python
