### 2024-08-02 dd78d3a
Removing printout of integral of g(r) from .gofr files
