### 2024-08-02 25e9651
Adding new getting started documentation for OpenMD
