+++ QUICK_START.md
@@ -1,0 +1,111 @@
+# Getting Started with OpenMD
+
+OpenMD is a simulation engine that is started from the command line, so we're going to assume that you are reasonably familiar with unix command line environments. This tutorial will focus on using OpenMD assuming it has already been installed on your machine. For directions on installing, checkout [INSTALL.md](docs/INSTALL.md) before continuing.
+
+Once OpenMD has been installed completely, and the `openmd` program is already in your path, you don't need to do anything special to set up your environment.  However, if you are just test-driving, and built it locally without installing it, you'll need to specify a few variables before continuing.  For example, if the code was unpacked into the directory `~/OpenMD-3.1` and you built in the subdirectory `~/OpenMD-3.1/build`, then you might need to do this:
+
+```bash
+export FORCE_PARAM_PATH="~/OpenMD-3.1/forceFields"
+export PATH="${PATH}:~/OpenMD-3.1/build/bin"
+```
+This will set your path to include the directory where openmd was built, and will reference the default force field files.  If you use `csh` or `tcsh` as your shell, the equivalent of those lines are:
+
+```bash
+setenv FORCE_PARAM_PATH ~/OpenMD-3.1/forceFields
+setenv PATH ${PATH}:~/OpenMD-3.1/build/bin
+```
+
+## Sample simulations
+
+In the `samples` subdirectory, there are a wide range of example simulations that have already been set up as demonstrations of the various capabilities of `OpenMD`.  We're going to start with a relatively simple case, a simulation of liquid water that uses the SPC/E water model for inter-atomic interactions.  Let's start by changing into the samples directory:
+
+```bash
+cd ~/OpenMD-3.1/samples
+```
+
+Now, let's go to the water/spce sample directory and run a simulation: 
+
+```bash
+cd water/spce
+openmd spce.omd                     # run the sample
+```
+
+As the simulations is progressing, you should see a progress bar indicating the time remaining in the simulation.
+If you compiled the parallel-aware version of OpenMD linked to MPI libraries, you can speed up the simulations by using multiple processors:
+
+```bash
+mpirun -np 4 openmd_MPI spce.omd   # run the sample, but with 4 processors
+```
+
+At the end of your simulation, you'll see a print out of a number of relevant thermodynamic quantities. You'll also notice a few new files in your sample directory, namely: 
+
+- `spce.dump` - the simulation trajectory file, containing multiple snapshots of atomic positions and velocities
+- `spce.eor` - the 'end-of-run' for the simulation, or the final configuration.
+- `spce.stat` - a 'statistics' file which records thermodynamically-interesting quantities like temperature, pressure, volume, potential, kinetic, and total energy
+- `spce.report` - statistical means and confidence intervals for the same quantities recorded in the `.stat` file
+
+In the next section, we'll explore our recommended ways of parsing and analyzing these file types.
+
+## Visualization Tools
+
+Suppose we want to see how the total, potential, and kinetic energies, the temperature, pressure, and volume of the box evolved as the simulation progressed. We can visualize the `.stat` file to look at these quantities:
+```bash
+xmgrace -nxy spce.stat         # visualize simulation properties as they evolve during the simulation
+```
+Here, we're  using the `xmgrace` package to plot our data, but `spce.stat` is just a tab-separated text file, so you can look at it in many different ways.  The first few lines of the `.stat` file give information about the version of the code that was used as well as the thermodynamic quantities that were recorded (and the units used).
+
+To visualize the trajectory or the contents of the `.dump` file after the simulation has completed we use a utility to generate an xyz file:
+```bash
[…]
