+++ CMakeLists.txt
@@ -1145,6 +1145,8 @@ install(
   FILES "${CMAKE_CURRENT_SOURCE_DIR}/docs/OpenMDmanual.pdf"
         "${CMAKE_CURRENT_SOURCE_DIR}/docs/AUTHORS"
         "${CMAKE_CURRENT_SOURCE_DIR}/docs/INSTALL.md"
+        "${CMAKE_CURRENT_SOURCE_DIR}/QUICK_START.md"
+        "${CMAKE_CURRENT_SOURCE_DIR}/README.md"
   DESTINATION docs
   PERMISSIONS OWNER_READ OWNER_WRITE GROUP_READ WORLD_READ)
 
