### 2024-08-02 #7004 @gezelter
Hi @samwaseda , we've now added a [QUICK_START.md](https://github.com/OpenMD/OpenMD/blob/main/QUICK_START.md) to the main directory which provides a step-by-step walk through of how to start and analyze one of the sample simulations.
