### 2024-08-09 #7004 @samwaseda
> Hello @samwaseda, a fresh OpenMD clone could be run using the following commands (assuming no dependencies and no installation):
>
> ```shell
> $ mkdir OpenMD && cd OpenMD
> $ git clone https://github.com/OpenMD/OpenMD.git openmd
> $ mkdir build && cd build
> $ cmake ../openmd/.
> $ make
> $ cd ../openmd/samples/argon
> $ set FORCE_PARAM_PATH=../../forceFields
> $ ../../../build/bin/openmd 500.omd
> ```
>
> This will allow you to run simulations locally on the most basic of systems. Some of the samples do require our dependencies and `MPI` to run in a reasonable time frame. More information can be found in our [INSTALL.md](https://github.com/OpenMD/OpenMD/blob/main/docs/INSTALL.md) file. Let me know if anything else is unclear or you're still unable to get OpenMD up and running.

Thanks for this snippet! This one is **very** helpful. I guess there's a small typo though: I presume there should be `export FORCE_PARAM_PATH=../../forceFields` and not `set FORCE_PARAM_PATH=../../forceFields` right?
