### 2024-08-09 #7004 @samwaseda
td;dr: I would like to see a very short `README.md` in every example with (1) what the example shows and (2) what to expect

> Hi @samwaseda , we've now added a [QUICK_START.md](https://github.com/OpenMD/OpenMD/blob/main/QUICK_START.md) to the main directory which provides a step-by-step walk through of how to start and analyze one of the sample simulations.

Thanks for this nice file! Yes, this was exactly the kind of document I wanted to see, or actually it's even more detailed than what I had imagined.

However, I have a bit of trouble connecting points in terms of documentation and actual code. Sorry for requesting a lot of things but I'm sure that this helps potential new users like me.

Now together with `QUICK_START.md` I see two more sources of documentation:
- [website](https://openmd.org)
- [pdf](https://github.com/OpenMD/OpenMD/blob/main/docs/OpenMDmanual.pdf) (this one is well written so I'd leave it like this, although I might love to see it in a more interactive format like html)

I can see they contain a huge amount of information, but as someone coming from outside, I would take a look at `README` or `QUICK_START` and try out the first examples, which is now possible with the new `QUICK_START.md`, but then when you go to the actual examples, the documentation is either totally missing, or the content varies largely from one example to another. Since one of the criteria for a JOSS paper includes automated tests, I would love to see what these examples do and what to expect for the output, so that users can know easily how to use it and whether they are using it correctly.

I'm not really sure what kind of purpose the website is supposed to fulfil, because the first posts (along with others) I see on the [example page](https://openmd.org/category/examples/) are actually not examples and then the examples do not seem to contain the full amount of information so that people can copy and paste the content to run them. If the page was not meant for new users to learn how to use OpenMD, I guess it's fine, but otherwise I think it's important to have a reference to a folder containing the all necessary files for the user to be able to run it.
