### 2024-08-09 #7004 @samwaseda
For the installation: Does something speak against including a step to update `$PATH`? I can imagine a lot of people want a local installation without root access.

Additionally, the list of dependencies should include the versions used for testing.
