### 2024-08-09 #7004 @crdrisko
Hello @samwaseda, thanks for your feedback so far. Yes, that `set`-command was a typo, I must have missed that in my conversions from csh to bash. As for the update to `$PATH`, there's really no reason to update that variable, other than the convenience of not having to carry around the relative or absolute path every time you want to run one of our applications. But the following:
```bash
$ openmd spce.omd
$ ~/path-to-openmd/build/bin/openmd spce.omd
```
are both equally valid. As one final example for this, when I'm developing/testing a new feature for OpenMD, I don't install or set the build path, because those changes may negatively impact our group, who all use the installed (latest version on main) applications. So, I carry around the relative path to my development build until the feature is verified, then I can go back to using the installed version. Hopefully that was clear, are you potentially looking for an explanation for something like that in the documentation?

To your other point, I will get working on the documentation updates for the individual sample directories and the results we expect from them. I'll update the thread when those changes go live.
