### 2024-08-09 07c2747
Starting documentation updates for the sample subdirectories
