+++ samples/oxides/README.md
@@ -1,4 +1,0 @@
-# Metal Oxides in OpenMD
-This Sample Library contains Metal Oxide compounds utilizing metals defined in the DR-EAM forcefield and oxygen.  This library contains a variety of compounds, both real and computed, that simulate multiple oxidation states for many of the metals.
-
-All .omd samples contained were obtained from CIF files found on [http://www.crystallography.net/cod/](http://www.crystallography.net/cod/).
