### 2024-08-09 #7004 @crdrisko
@samwaseda, as a follow up to my post this morning, we've made some progress on implementing your suggested changes to documentation in the sample subdirectories. I've included a template README file that all samples should use, and the `samples/air`, `samples/argon`, and `samples/builders` now all follow this format. I'll continue working on it over the coming days, but hopefully this was the direction you were hoping for.
