### 2024-08-10 #7004 @samwaseda
So, I carry around the relative path to my development build until the feature is verified, then I can go back to using the installed version. Hopefully that was clear, are you potentially looking for an explanation for something like that in the documentation?
>

Not necessarily. It just feels to me that with conda and other package managers the tendency is that everything works out of the box so I thought it might be confusing when it might not work depending on the example, but obviously whoever has the slightest understanding of bash they would know how to make it work so if you consider it unnecessary I'm also fine with that.
