### 2024-08-10 #7004 @samwaseda
> @samwaseda, as a follow up to my post this morning, we've made some progress on implementing your suggested changes to documentation in the sample subdirectories. I've included a template README file that all samples should use, and the `samples/air`, `samples/argon`, and `samples/builders` now all follow this format. I'll continue working on it over the coming days, but hopefully this was the direction you were hoping for.

I took a look and yes it's exactly the kind of documentation I was hoping for. I also like the length of the documents.
