### 2024-08-10 04e0ee7
Update INSTALL.md
