### 2024-08-10 ed048f0
Update INSTALL.md
