### 2024-08-10 #7004 @gezelter
> For the installation: Does something speak against including a step to update `$PATH`? I can imagine a lot of people want a local installation without root access.

Hi @samwaseda , we've added $PATH information in [INSTALL.md](https://github.com/OpenMD/OpenMD/blob/main/docs/INSTALL.md)  and in QUICK_START.md(https://github.com/OpenMD/OpenMD/blob/main/QUICK_START.md)

>
> Additionally, the list of dependencies should include the versions used for testing.

This has now been added in both  [INSTALL.md](https://github.com/OpenMD/OpenMD/blob/main/docs/INSTALL.md)  and [README.md](https://github.com/OpenMD/OpenMD/blob/main/README.md)
