+++ samples/graphene/README.md
@@ -1,3 +1,11 @@
+# Graphene
+
+## Background Information
+
+## Instructions
+
+## Expected Output
+
 This sample illustrates a simulation of united-atom (UA) propylene
 monomers confined between and all-atom (AA) representation of graphene
 sheets.  The force field is based partially on OPLS-AA, and partially
