+++ samples/DR-EAM/README.md
@@ -1,0 +1,7 @@
+# Density Readjusting Embedded Atom Method (DR-EAM)
+
+## Background Information
+
+## Instructions
+
+## Expected Output
