+++ samples/Madelung/README.md
@@ -1,8 +1,15 @@
-# Madelung Energy Sample Calculation
+# Madelung Energy
+
+## Background Information
+
+## Instructions
+
+## Expected Output
 
 The sample in this directory provides a way of checking the value of
 the Madelung Energy for a perfect crystal of NaCl.  The relevant
 quantities are:
+
 ```
 M (Madelung constant) = 1.74756
 a (lattice constant)  = 5.65 Angstroms
@@ -12,6 +19,7 @@ M q^2 / (4 pi e0 a)   = 102.708173 kcal / mol
 
 The file NaCl.omd contains 8000 ions, so the total electrostatic energy
 of the perfect crystal in this file should be:
+
 ```
 V_electrostatic = -821665.38  kcal / mol
 ```
@@ -34,6 +42,7 @@ V_electrostatic = -821667.68 kcal / mol
 
 To obtain values for the electrostatic potential in OpenMD, we add the
 `ELECTROSTATIC_POTENTIAL` keyword to the end of the statFileFormat:
+
 ```
 statFileFormat = "TIME|TOTAL_ENERGY|POTENTIAL_ENERGY|KINETIC_ENERGY|TEMPERATURE|PRESSURE|VOLUME|CONSERVED_QUANTITY|ELECTROSTATIC_POTENTIAL";
 ```
