+++ samples/zeolite/README.md
@@ -1,4 +1,6 @@
-# Sample zeolite structures
+# Zeolite
+
+## Background Information
 
 OpenMD supports the **CLAYFF** forcefield described in: "Molecular
 Models of Hydroxide, Oxyhydroxide, and Clay Phases and the Development
@@ -6,3 +8,7 @@ of a General Force Field," Randall T. Cygan, Jian-Jie Liang, and
 Andrey G. Kalinichev, *J. Phys. Chem. B* **108**, pp. 1255-1266 (2004)
 DOI: [10.1021/jp0363287](https://doi.org/10.1021/jp0363287), and
 provides an example of a solvated ZSM-5 zeolite structure.
+
+## Instructions
+
+## Expected Output
