+++ samples/bond-order/README.md
@@ -2,28 +2,92 @@
 
 ## Background Information
 
-These are sample files used to test the bond order parameter
-code in StaticProps.  Perfect SC, BCC, FCC, HCP, and Icosahedral
-clusters have known analytic values for the bond order parameter
-and the files in this directory can be used to make sure the
-correct values are being computed.  The central atom in each
-cluster is a copper atom, so the proper way to run the tests would
-be a command like:
+One way to study the local environments around atoms is the use of Bond-orientational order parameters. These were first presented in 
 
-  $(OPENMD_HOME)/bin/StaticProps --bo -i bcc.omd --rcut=9 --sele1="select Cu"
+> "Bond-orientaional order in liquids and glasses," P. J. Steinhart, 
+> D. R. Nelson, and M. Ronchetti, Phys. Rev. B, 28, 784 (1983).
 
-legend:
+and formulae for individual atoms appeared in:
 
-  hcp.omd                 = Hexagonal Close Packed structure
-  bcc.omd                 = Body Centered Cubic structure
-  fcc.omd                 = Face Centered Cubic structure
-  sc.omd                  = Simple Cubic structure
-  icosahedron.omd         = Icosahedral cluster
-  surface_icosahedron.omd = Half of a Icosahedral Cluster approximating a nanoparticle surface atom
-  tet.omd                 = Tetrahedral cluster
+> "Numerical calculation of the rate of crystal nucleation in a 
+> Lennard-Jones system at moderate undercooling," Pieter Rein ten Wolde, 
+> Maria J. Ruiz-Montero, and Daan Frenkel, J. Chem. Phys. 104, pp. 
+> 9932-9947 (1996).
 
-(Cu atoms are located at (0,0,0), and Au atoms surround Cu in the clusters.)
+`StaticProps` can compute bond orientational order parameters, using a single cutoff radius to decide membership in the list of neighbors, and does not have use a distance-dependent weighting as used in the second reference above.
 
 ## Instructions
 
+This directory contains sample files used to test the orientational bond 
+order parameter code in `StaticProps`.  Perfect SC, BCC, FCC, HCP, and 
+Icosahedral clusters have known analytic values for the bond order parameter and the files in this directory can be used to make sure the correct values are being computed.  The central atom in each cluster is a copper atom, so the proper way to run the tests would be a command like:
+
+```
+StaticProps --bo -i bcc.omd --rcut=9 --sele1="select Cu"
+```
+
+The selection script will look at specific types of central atoms - in this case, copper, and dynamic selections can also be utilized, e.g. `--sele1="within(5, Au)"`.  By default, this command computes the $Q_{l}$ and $\hat{W}_{l}$ parameters up to $l = 12$.  The configurational averages of these values as well as the distributions of atomic $q_{l}$ and $\hat{w}_{l}$  values are then placed in `.boq` and `.bow` files.
+
+The `bcc.boq` file that is generated by the command above should look like (only the header is shown):
+```
+# Bond Order Parameters
+# selection script: "select Cu"
+# parameters:  rcut = 9, nbins = 0
+# <Q_0>: 1
+# <Q_1>: 3.37141e-17
+# <Q_2>: 5.54483e-17
+# <Q_3>: 2.99407e-17
+# <Q_4>: 0.0363696
+# <Q_5>: 8.02427e-17
+# <Q_6>: 0.510688
+# <Q_7>: 2.17378e-16
[…]
