+++ samples/bond-order/README.md
@@ -2,33 +2,50 @@
 
 ## Background Information
 
-One way to study the local environments around atoms is the use of Bond-orientational order parameters. These were first presented in 
+One way to study the local environments around atoms is the use of
+Bond-orientational order parameters. These were first presented in
 
-> "Bond-orientaional order in liquids and glasses," P. J. Steinhart, 
+> "Bond-orientaional order in liquids and glasses," P. J. Steinhart,
 > D. R. Nelson, and M. Ronchetti, Phys. Rev. B, 28, 784 (1983).
 
 and formulae for individual atoms appeared in:
 
-> "Numerical calculation of the rate of crystal nucleation in a 
-> Lennard-Jones system at moderate undercooling," Pieter Rein ten Wolde, 
-> Maria J. Ruiz-Montero, and Daan Frenkel, J. Chem. Phys. 104, pp. 
-> 9932-9947 (1996).
+> "Numerical calculation of the rate of crystal nucleation in a
+> Lennard-Jones system at moderate undercooling," Pieter Rein ten
+> Wolde, Maria J. Ruiz-Montero, and Daan Frenkel, J. Chem. Phys. 104,
+> pp.  9932-9947 (1996).
 
-`StaticProps` can compute bond orientational order parameters, using a single cutoff radius to decide membership in the list of neighbors, and does not have use a distance-dependent weighting as used in the second reference above.
+`StaticProps` can compute bond orientational order parameters, using a
+single cutoff radius to decide membership in the list of neighbors,
+and does not have use a distance-dependent weighting as used in the
+second reference above.
 
 ## Instructions
 
-This directory contains sample files used to test the orientational bond 
-order parameter code in `StaticProps`.  Perfect SC, BCC, FCC, HCP, and 
-Icosahedral clusters have known analytic values for the bond order parameter and the files in this directory can be used to make sure the correct values are being computed.  The central atom in each cluster is a copper atom, so the proper way to run the tests would be a command like:
+This directory contains sample files used to test the orientational
+bond order parameter code in `StaticProps`.  Perfect simple cubic
+(SC), body-centered cubic (BCC), face-centered cubic (FCC), hexagonal
+close packed (HCP), and Icosahedral clusters have known analytic
+values for the bond order parameter and the files in this directory
+can be used to make sure the correct values are being computed.  The
+central atom in each cluster is a copper atom, so the proper way to
+run the tests would be a command like:
 
 ```
 StaticProps --bo -i bcc.omd --rcut=9 --sele1="select Cu"
 ```
 
-The selection script will look at specific types of central atoms - in this case, copper, and dynamic selections can also be utilized, e.g. `--sele1="within(5, Au)"`.  By default, this command computes the $Q_{l}$ and $\hat{W}_{l}$ parameters up to $l = 12$.  The configurational averages of these values as well as the distributions of atomic $q_{l}$ and $\hat{w}_{l}$  values are then placed in `.boq` and `.bow` files.
+The selection script will look at specific types of central atoms - in
+this case, copper, and dynamic selections can also be utilized,
+e.g. `--sele1="within(5, Au)"`.  By default, this command computes the
+$Q_{l}$ and $\hat{W}_{l}$ parameters up to $l = 12$.  The
+configurational averages of these values as well as the distributions
[…]
