### 2024-08-12 fe53685
Completing the aqueous ion sample readme
