### 2024-08-13 0db7b57
Making Madelung README more readable.
