+++ samples/Madelung/README.md
@@ -1,33 +1,32 @@
 # Madelung Energy
 
 ## Background Information
-
-## Instructions
-
-## Expected Output
-
 The sample in this directory provides a way of checking the value of
-the Madelung Energy for a perfect crystal of NaCl.  The relevant
-quantities are:
+the Madelung Energy for a perfect crystal of NaCl using different methods of computing electrostatic interactions.  The relevant quantities are:
 
-```
-M (Madelung constant) = 1.74756
-a (lattice constant)  = 5.65 Angstroms
-q^2 / (4 pi e0 a)     = 58.77233   kcal / mol
-M q^2 / (4 pi e0 a)   = 102.708173 kcal / mol
-```
+$$
+\begin{aligned}
+M \text{ (Madelung constant)} &= 1.74756 \\
+a \text{ (lattice constant)} &= 5.65 \AA \\
 
-The file NaCl.omd contains 8000 ions, so the total electrostatic energy
+\frac{q^2}{4 \pi \epsilon_0 a} &= 58.77233 \text{ kcal mol}^{-1} \\
+M \frac{q^2}{4 \pi \epsilon_0 a} & = 102.708173 \text{ kcal mol}^{-1}\\
+\end{aligned}
+$$ 
+
+The file `NaCl.omd` contains 8000 ions, so the total electrostatic energy
 of the perfect crystal in this file should be:
 
-```
-V_electrostatic = -821665.38  kcal / mol
-```
+$$
+V_\text{electrostatic} = -821665.38 \text{ kcal mol}^{-1}
+$$
 
+## Instructions
 Using different electrostatic calculation methods, we can get quite
 close to this value.
 
-For example, with :
+For example, to test the Damped Shifted Force (DSF) model with a very large cutoff, we set these lines in `NaCl.omd` :
+
 ```
 cutoffMethod = "shifted_force";
 electrostaticScreeningMethod = "damped";
@@ -35,14 +34,41 @@ cutoffRadius = 28;
 dampingAlpha = 0.14159292;
 ```
 
-The resultant electrostatic potential is:
-```
-V_electrostatic = -821667.68 kcal / mol
-```
+Other options for the `cutoffMethod` include `shifted_potential`, `reaction_field`, `ewald_full`, and `hard`, although `shifted_force` provides remarkable accuracy at minimal computational cost.
 
-To obtain values for the electrostatic potential in OpenMD, we add the
-`ELECTROSTATIC_POTENTIAL` keyword to the end of the statFileFormat:
+To extract values for the electrostatic potential in OpenMD, we add the
+`ELECTROSTATIC_POTENTIAL` keyword to the end of the `statFileFormat`:
 
 ```
 statFileFormat = "TIME|TOTAL_ENERGY|POTENTIAL_ENERGY|KINETIC_ENERGY|TEMPERATURE|PRESSURE|VOLUME|CONSERVED_QUANTITY|ELECTROSTATIC_POTENTIAL";
 ```
+
+Once those lines have been set we run a single step simulation:
+
+```
+mpirun -np 4 openmd_MPI NaCl.omd
+```
+
+## Expected Output
+The report generated from this simulation should look like:
+```
+###############################################################################
+# Status Report:                                                              #
+#              Total Time:           2 fs                                     #
+#       Number of Samples:           2                                        #
+#            Total Energy:     -709366  ±  0            kcal/mol              #
+#        Potential Energy:     -709366  ±  0            kcal/mol              #
+#          Kinetic Energy: 2.34108e-25  ±  3.24457e-25  kcal/mol              #
+#             Temperature: 9.81846e-27  ±  1.36077e-26  K                     #
+#                Pressure:      116088  ±  0.00118777   atm                   #
[…]
