### 2024-08-13 9210d8a
Update README.md for equations
