+++ samples/Madelung/README.md
@@ -4,22 +4,17 @@
 The sample in this directory provides a way of checking the value of
 the Madelung Energy for a perfect crystal of NaCl using different methods of computing electrostatic interactions.  The relevant quantities are:
 
-$$
-\begin{aligned}
+$$\begin{aligned}
 M \text{ (Madelung constant)} &= 1.74756 \\
 a \text{ (lattice constant)} &= 5.65 \AA \\
-
 \frac{q^2}{4 \pi \epsilon_0 a} &= 58.77233 \text{ kcal mol}^{-1} \\
 M \frac{q^2}{4 \pi \epsilon_0 a} & = 102.708173 \text{ kcal mol}^{-1}\\
-\end{aligned}
-$$ 
+\end{aligned}$$ 
 
 The file `NaCl.omd` contains 8000 ions, so the total electrostatic energy
 of the perfect crystal in this file should be:
 
-$$
-V_\text{electrostatic} = -821665.38 \text{ kcal mol}^{-1}
-$$
+$$V_\text{electrostatic} = -821665.38 \text{ kcal mol}^{-1}$$
 
 ## Instructions
 Using different electrostatic calculation methods, we can get quite
@@ -68,7 +63,7 @@ The report generated from this simulation should look like:
 ```
 
 Looking into the `NaCl.stat` file shows that the resultant electrostatic potential is:
-$$
-V_\text{electrostatic} = -821667.61 \text{ kcal mol}^{-1}
-$$
+
+$$V_\text{electrostatic} = -821667.61 \text{ kcal mol}^{-1}$$
+
 which is only a 0.0002714 % error from the true Madelung energy.
