### 2024-08-13 6351158
Updated samples for electrostatics
