+++ samples/Madelung/dipoles/README.md
@@ -1,30 +1,44 @@
+# Dipolar Crystals
+
+## Background Information
 This directory contains a set of dipolar crystals that can be used to
 test electrostatic energies for dipole-dipole interactions.  The
 dipolar analogues to the Madelung constants for ionic crystals were
 first worked out by Sauer who computed the energies of certain
 selected dipole arrays (ordered arrays of zero magnetization) and
-obtained a number of these constants.[1]
+obtained a number of these constants:
 
-This theory was developed more completely by Luttinger & Tisza [2] who
+> J. A. Sauer, "Magnetic Energy Constants of Dipolar Lattices,"
+> Phys. Rev. 57, 142–146 (1940) doi: 10.1103/PhysRev.57.142
+
+
+This theory was developed more completely by Luttinger & Tisza who
 tabulated energy constants for the Sauer arrays (and other periodic
-structures).  We have repeated the Luttinger & Tisza series summations
-to much higher order and obtained the following energy constants:
+structures):
 
-Array Type   Lattice    Dipole Direction     Energy constants
-----------   -------    ----------------     ----------------
-   A           SC             001               -2.676788684
-   A           BCC            001                0
-   A           BCC            111               -1.770078733
-   A           FCC            001                2.166932835
-   A           FCC            011               -1.083466417
+> J. M. Luttinger and L. Tisza, "Theory of Dipole Interaction in
+> Crystals," Phys. Rev. 70, 954–964 (1946) doi: 10.1103/PhysRev.70.954
+> Also note the errata contained in: Phys. Rev. 72, 257 (1947) 
+> doi: 10.1103/PhysRev.72.257
 
-   *           BCC          minimum             -1.985920929
+We have repeated the Luttinger & Tisza series summations to higher
+order and obtained the following energy constants:
 
-   B           SC             001               -2.676788684
-   B           BCC            001               -1.338394342
-   B           BCC            111               -1.770078733
-   B           FCC            001               -1.083466417
-   B           FCC            011               -1.807573634
+| Array Type | Lattice | Dipole Direction | Energy constants |
+|------------|---------|------------------|------------------|
+|   A        |   SC    |         001      |   -2.676788684   |
+|   A        |   BCC   |         001      |    0             |
+|   A        |   BCC   |         111      |   -1.770078733   |
+|   A        |   FCC   |         001      |    2.166932835   |
+|   A        |   FCC   |         011      |   -1.083466417   |
+|            |         |                  |                  |
+|   *        |   BCC   |       minimum    |   -1.985920929   |
+|            |         |                  |                  |
+|   B        |   SC    |         001      |   -2.676788684   |
+|   B        |   BCC   |         001      |   -1.338394342   |
+|   B        |   BCC   |         111      |   -1.770078733   |
+|   B        |   FCC   |         001      |   -1.083466417   |
+|   B        |   FCC   |         011      |   -1.807573634   |
  
 Type "A" arrays have nearest neighbor strings of antiparallel dipoles.
 
@@ -33,46 +47,103 @@ if the dipoles are contained in a plane perpendicular to the dipole
 direction that passes through the dipole.
 
 There's also an additional minimum energy structure for the BCC
-lattice that was found by Luttinger & Tisza.  All of the energy
+lattice that was found by Luttinger & Tisza. All of the energy
[…]
