### 2024-08-13 0c27ac2
Rolling back electrostatics change and implementing a per-frame sanity check
