+++ src/applications/staticProps/BondOrderParameter.hpp
@@ -87,7 +87,7 @@ namespace OpenMD {
   public:
     BondOrderParameter(SimInfo* info, const std::string& filename,
                        const std::string& sele, double rCut,
-		       unsigned int nbins);
+                       unsigned int nbins);
 
     virtual void process();
 
