### 2024-08-14 6acd4c1
Updating dipolar and quadrupolar example python scripts
