### 2024-08-14 39032ec
gofr comparison in aqueous ion sample
