### 2024-08-14 b264c61
Adding README for zeolite / water sample
