+++ samples/zeolite/README.md
@@ -44,7 +44,8 @@ xmgrace zsm5_loaded.rcorr
 
 To calculate diffusivity, we need the slope of the mean squared
 displacement (rcorr) that was generated by `DynamicProps`,
-$$ D = \lim_{t \rightarrow \infty} \frac{1}{6 t} \langle | r(t) - r(0)
+
+$$D = \lim_{t \rightarrow \infty} \frac{1}{6 t} \langle | r(t) - r(0)
 |^2 \rangle $$
 
 ## Expected Output
