+++ samples/graphene/README.md
@@ -2,20 +2,18 @@
 
 ## Background Information
 
-## Instructions
-
-## Expected Output
-
 This sample illustrates a simulation of united-atom (UA) propylene
 monomers confined between and all-atom (AA) representation of graphene
 sheets.  The force field is based partially on OPLS-AA, and partially
 on TraPPE-UA.
 
-Propylene parameters from: C.D. Wick, M.G. Martin, and J.I. Siepmann,
-"Transferable potentials for phase equilibria. 4. United-atom
-description of linear and branched alkenes and of alkylbenzenes,"
-*J. Phys. Chem. B*, **104**, pp. 8008-8016 (2000).
-DOI: [10.1021/jp001044x](https://doi.org/10.1021/jp001044x)
+Propylene parameters are from: 
+
+> C.D. Wick, M.G. Martin, and J.I. Siepmann, "Transferable potentials
+> for phase equilibria. 4. United-atom description of linear and
+> branched alkenes and of alkylbenzenes," *J. Phys. Chem. B*, **104**,
+> pp. 8008-8016 (2000). DOI:
+> [10.1021/jp001044x](https://doi.org/10.1021/jp001044x)
 
 Note that the molecule definition in `graphene.inc` includes bonds that span the
 box boundaries.  If you want to extend to larger boxes, start with the molecule
@@ -37,6 +35,8 @@ Files included:
    input script that places propylene molecules inside bounds of the
    graphene sheets.
 
+## Instructions
+
 To create the initial configurations, we typically run:
 
 `packmol < system.pack`
@@ -47,25 +47,28 @@ To get this into a reasonable OpenMD starting configuration, we would run:
 `atom2omd -ixyz system.xyz`
 
 This creates `system.omd`, which has periodic box guessed from the bounding box
-of the solvent molecules, and this box is generally not the same size as the
+of the solvent molecules, and this box is generally **not** the same size as the
 box containing the graphene sheets.  To successfully combine the monomers with
 the graphene sheets, the `system.omd` file must be edited to modify the Hmat
 line to read:
 
-~~~~
+```
         Hmat: {{ 22.23, 0, 0 }, { 0, 42.78, 0 }, { 0, 0, 50 }}
-~~~~
+```
 
 Then to combine the propylene monomers with the graphene sheets, eliminating
 molecules that overlap, we would run:
 
-`omd-solvator -u graphene.omd -v system.omd -r 3.5 -o combined.omd -n 360 -p 3`
+```
+omd-solvator -u graphene.omd -v system.omd -r 3.5 -o combined.omd -n
+360 -p 3
+```
 
 Following this, we typically edit the `combined.omd` file to include the
 molecule definitions, set the force field, and set various simulation
 parameters:
 
-~~~~
+```
 #include "graphene.inc"
 #include "propylene.omd"
 
@@ -75,7 +78,7 @@ component{
 }
 component{
   type = propylene;
-nMol = 23;
+nMol = 27;
 }
 
 forceField = "graphene";
@@ -87,7 +90,61 @@ dampingAlpha = 0.2;
 targetTemp = 300;
 tauThermostat = 1000;
 dt = 1.0;
-runTime = 1e3;
-sampleTime = 100;
+runTime = 1e5;
+sampleTime = 1000;
 statusTime = 10;
-~~~~
+```
+
+Once these modifications are in place, we can sample some initial velocities and run the simulation:
+
+```
+thermalizer -t 300 -i combined.omd -o warm.omd
+mpirun -np 4 openmd_MPI combined.omd
+```
+
+## Expected Output
+
+The report from this simulation should look similar to:
+```
+###############################################################################
+# Status Report:                                                              #
+#              Total Time:      100000 fs                                     #
+#       Number of Samples:       10001                                        #
[…]
