+++ samples/RNEMD/README.md
@@ -2,7 +2,42 @@
 
 ## Background Information
 
+There are many ways to compute transport properties from molecular
+dynamics simulations.  Equilibrium Molecular Dynamics (EMD)
+simulations can be used by computing relevant time correlation
+functions and assuming linear response theory holds.  For some transport
+properties (notably thermal conductivity), EMD approaches
+are subject to noise and poor convergence of the relevant
+correlation functions. Traditional Non-equilibrium Molecular Dynamics
+(NEMD) methods impose a gradient (e.g. thermal or momentum) on a
+simulation.  However, the resulting flux is often difficult to
+measure. Furthermore, problems arise for NEMD simulations of
+heterogeneous systems, such as phase-phase boundaries or interfaces,
+where the type of gradient to enforce at the boundary between
+materials is unclear.
+
+*Reverse* Non-Equilibrium Molecular Dynamics (RNEMD) methods adopt
+a different approach in that an unphysical *flux* is imposed between
+different regions or "slabs" of the simulation box.  The goal of RNEMD
+methods is to calculate the relevant transport property ($\lambda$)
+that connects the flux ($\textbf{J}$) and driving force ($\nabla X$)
+according to the generalized equation,
+
+$$
+\textbf{J} = - \lambda \nabla X.
+$$
+
+Since the amount of the applied flux is known exactly, and the measurement of
+gradient is generally less complicated, imposed-flux methods typically take
+shorter simulation times to obtain converged results for transport properties.
+
 ## Instructions
+
+With all examples mentioned below, you can run `openmd` with the following command:
+
+```
+$ mpirun -np 4 openmd_MPI <sample_file>.omd
+```
 
 ### Example 1
 
@@ -11,8 +46,11 @@ which has the momentum flux functionality of RNEMD turned on. Notice
 in the RNEMD block of the (.omd) file that
 
 ```C++
-	fluxType = "Px";
-	momentumFlux = 6.0e-7;
+RNEMD {
+  fluxType = "Px";
+  momentumFlux = 6.0e-7;
+  // ...
+}
 ```
 
 With these parameters, the applied flux will be an x-axis momentum
@@ -25,16 +63,19 @@ flux to the system's gradient response of the velocity.
 
 It is also possible to measure the *thermal conductivity* of a
 material using the RNEMD functionality in OpenMD. As an example,
-**graphene.omd** is an (.omd) file where two sheets of graphene have a
+`graphene.omd` is an (.omd) file where two sheets of graphene have a
 thermal flux applied accross the long axis of the sheet. In the (.omd)
 file the fluxType has been set to a kinetic energy flux, and also that
 the kineticFlux is defined.
 
 ```C++
-	fluxType = "KE";
-	kineticFlux = -6.55e-11;
+RNEMD {
+  fluxType = "KE";
+  kineticFlux = -6.55e-11;
+  // ...
+}
 ```
-	
+
 The system responds to the thermal flux by developing a temperature
 gradient across the z-axis of the system. The thermal conductivity can
 be computed by relating the resulting thermal gradient to the imposed
@@ -50,13 +91,16 @@ OpenMD easily allows for this, and the RNEMD definition block in the
 only one component, now we need to allow for more than one atom or
 molecule type to be selected.
 
-**gold_water_interface.omd** is an example system where we can compute
+`gold_water_interface.omd` is an example system where we can compute
 the thermal conductivity across an interface, here, a gold / water
 interface. Notice in the RNEMD declaration block in the (.omd) file
[…]
