### 2024-08-15 1a1c5d4
New section for next steps added to QUICK_START
