+++ samples/nanoparticles/README.md
@@ -1,0 +1,138 @@
+# Metallic nanoparticle Samples
+
+
+## Background Information
+
+In these examples, we’ll build gold nanoparticles and equilibrate them to a temperature of 300K. 
+
+## Instructions
+### Spherical metal nanoparticle
+
+1. We need to start with an input `.omd` file for the metal of your choice. This directory has a `gold.omd` for this example. It looks like:
+```
+<OpenMD>
+  <MetaData>
+molecule{
+  name = "Au";
+
+  atom[0]{
+    type = "Au";
+    position(0.0, 0.0, 0.0);
+  }
+}
+
+component{
+  type = "Au";
+  nMol = 1;
+}
+
+forceField = "SC";
+forceFieldFileName = "SuttonChen.QSC.frc";
+  </MetaData>
+</OpenMD>
+```
+2. Now to build the nanoparticle. We’ll choose a radius of 15 Å and use the lattice constant of gold, 4.08 Å. We’ll call our initial structure NP15.omd. 
+```
+nanoparticleBuilder --latticeConstant=4.08 --radius=15 gold.omd -o NP15.omd
+```
+
+3. To take a look at the structure we’ve just created, you can use the following command:
+```
+Dump2XYZ -i NP15.omd
+```
+to create a file called `NP15.xyz`, which can be viewed in VMD, Jmol, or any other chemical structure viewer.
+
+4. Add the following lines to the new NP15.omd file below the `forceFieldFileName` line. This sets some details for the simulation. 
+```
+ensemble = "LHull";
+targetTemp = 5;
+targetPressure = 1;
+viscosity = 0.1;
+dt = 4.0;
+runTime = 2E5;
+
+sampleTime = 2000.0;
+statusTime = 4;
+seed = 985456376;
+
+usePeriodicBoundaryConditions = "false";
+tauThermostat = 1E3;
+tauBarostat = 5E3;
+```
+
+5. NanoparticleBuilder carves a nanoparticle of our chosen radius out of a perfect gold crystal. We need to give the atoms some initial velocities before we start equilibrating. We’ll start it out at 5K:
+```
+thermalizer -t 5 NP15.omd -o NP15_5K.omd
+```
+6. For the first step in the equilibration we need to let the gold lattice structurally relax. `NP15_5K.omd` can now be run:
+```
+openmd NP15_5K.omd
+```
+7. Running the simulation will create several new files. `NP15_5K.dump` contains the trajectory of the simulation. Statistics such as temperature, pressure, and energy will be recorded in the `NP15_5K.stat` file and can be viewed using:
+```
+xmgrace -nxy NP15_5K.stat
+```
+8. The end-of-run file `NP15_5K.eor` stores the last configuration of the simulation. We’ll copy it to a new .omd file.
+```
+cp NP15_5K.eor NP15_100K.omd
+```
+9. To continue with the equilibration we need to change the `targetTemp` of `NP15_100K.omd`. We’ll increase it to 100 and run the `NP15_100K.omd` file.
+10. We’ll continue the procedure of copying the .eor file to a new .omd file and increasing the temperature until we’ve reached 300K. Temperature increases of 50 – 100K and simulation times of 100 – 200 ps are reasonable.
+
+### Icosahedral metal nanoparticle
+
+1. To start the process, a metal file (`gold.omd`) is needed to describe the material composition of the particle.  We'll use the same file from the spherical example above.
+
+2. To create the particle coordinates, use the command:
+```
+icosahedralBuilder -o file.omd --shell=8 --latticeConstant=4.08 --ico gold.omd
+```
+where `file.omd` is the name of your icosahedra, 8 is the number of shells (in this example), 4.08 is the lattice constant in Angstroms for the metal (4.08 is the correct value for gold), and the skeletal OpenMD file above is called `gold.omd`
+
[…]
