+++ samples/DR-EAM/README.md
@@ -1,7 +1,131 @@
 # Density Readjusting Embedded Atom Method (DR-EAM)
+The samples in this directory illustrate a new method for simulating
+bulk metals while treating polarizability.
 
 ## Background Information
+The Density Readjusting Embedded Atom Method (DR-EAM):
+
+> Bhattarai, H., Newman, K. E., & Gezelter, J. D. (2019). Polarizable potentials for metals:
+> The density readjusting embedded atom method (DR-EAM). Physical Review
+> B, 99(9), 94106, doi: 10.1103/PhysRevB.99.094106
+
+allows fluctuating densities within the framework of the Embedded Atom
+Method (EAM):
+
+> Daw, M. S., & Baskes, M. I. (1984). Embedded-atom method: Derivation
+> and application to impurities, surfaces, and other defects in
+> metals. Physical Review B, 29(12), 6443–6453. doi:
+> 10.1103/PhysRevB.29.6443
+
+by adding an additional degree of freedom, the charge for each atomic
+site.  The samples here illustrate the use of DR-EAM in bulk metals,
+surfaces, and alloys.
 
 ## Instructions
 
+### Ordered Structures
+
+In the `orderedStructures`  directory, are two alloy samples: Ag<sub>3</sub>Au in
+the $L1_{2}$ structure and PtNi in the $L1_{0}$ structure.  Both of the `omd` files in this directory describe structure optimizations, and not molecular dynamics. The `omd` files have a block to use Conjugate Gradient `CG` as the method for finding the optimal structure:
+
+```
+minimizer {
+  useMinimizer = True;
+  method = CG;
+  maxIterations = 5000;
+}
+```
+These files also turn on the fluctuating charge module to use DR-EAM.  Here, the target temperature for the electronic degrees of freedom is set very low (1K):
+```
+flucQ {
+  propagator = "Langevin";
+  targetTemp = 1;
+  dragCoefficient = 0.1;
+}
+```
+
+To carry out these optimization runs, we would use these commands:
+
+```
+openmd AgAuL12.omd
+openmd PtNiL10.omd
+```
+
+OpenMD will simultaneously optimize the structure and the partial charges on the atoms.  Once the structures have been optimized, we can look at the charges (`q`) that have been assigned to the types of atoms at the end of the run (in the `eor` file):
+
+```
+StaticProps -i AgAuL12.eor -q --sele1="select Ag_FQ" -o Ag.chargeHist
+StaticProps -i AgAuL12.eor -q --sele1="select Au_FQ" -o Au.chargeHist
+StaticProps -i PtNiL10.eor -q --sele1="select Pt_FQ" -o Pt.chargeHist
+StaticProps -i PtNiL10.eor -q --sele1="select Ni_FQ" -o Ni.chargeHist
+```
+These calculate the distribution of partial charges assigned to particular types of atoms.  Plotting these:
+
+```
+xmgrace Ag.chargeHist Au.chargeHist Pt.chargeHist Ni.chargeHist
+```
+we should expect to see Ag and Au with opposing charges (but in a 3:1 ratio), and Pt and Ni with opposing charges.
+
+### Field Screening
+
+One of the properties of a metal is the ability to screen atoms internal to the slab from external electric fields. The sample in the `fieldScreening` subdirectory  is a Platinum slab simulated in the presence of an external electric field.  The relevant portions of that sample file add a few new items to the simulation definition:
+```
+flucQ {
+  propagator = "Langevin";
+  targetTemp = 1;
+  dragCoefficient = 0.1;
+}
+outputFluctuatingCharges = true;   # we want to see the charges layer-by-layer
+outputElectricField = true;        # we want to measure the field inside the slab
+outputDensity = true;              # helps us look at the electron density
[…]

[…]
