### 2024-08-16 a22b876
Adding protein sample
