### 2024-08-16 #7004 @samwaseda
From my side it's almost good to go. There's just "state of the field" missing in the paper, which is one checkbox in the [list above](https://github.com/openjournals/joss-reviews/issues/7004#issuecomment-2263006845). I guess there's no mentioning of other software packages in the current version? I'm not a big MD person but I would say at least LAMMPS should be there.
