+++ docs/paper/paper.bib
@@ -10,8 +10,6 @@
 
 @article{OpenMM,
 	author = {Eastman, Peter AND Swails, Jason AND Chodera, John D. AND McGibbon, Robert T. AND Zhao, Yutong AND Beauchamp, Kyle A. AND Wang, Lee-Ping AND Simmonett, Andrew C. AND Harrigan, Matthew P. AND Stern, Chaya D. AND Wiewiora, Rafal P. AND Brooks, Bernard R. AND Pande, Vijay S.},
-	date-added = {2024-08-16 10:26:15 -0400},
-	date-modified = {2024-08-16 10:26:19 -0400},
 	doi = {10.1371/journal.pcbi.1005659},
 	journal = {PLOS Computational Biology},
 	month = {07},
@@ -19,48 +17,34 @@
 	pages = {1-17},
 	publisher = {Public Library of Science},
 	title = {{OpenMM} 7: Rapid development of high performance algorithms for molecular dynamics},
-	url = {https://doi.org/10.1371/journal.pcbi.1005659},
 	volume = {13},
-	year = {2017},
-	bdsk-url-1 = {https://doi.org/10.1371/journal.pcbi.1005659}}
+	year = {2017}}
 
 @article{CHARMM,
 	author = {Brooks, B. R. and Brooks III, C. L. and Mackerell Jr., A. D. and Nilsson, L. and Petrella, R. J. and Roux, B. and Won, Y. and Archontis, G. and Bartels, C. and Boresch, S. and Caflisch, A. and Caves, L. and Cui, Q. and Dinner, A. R. and Feig, M. and Fischer, S. and Gao, J. and Hodoscek, M. and Im, W. and Kuczera, K. and Lazaridis, T. and Ma, J. and Ovchinnikov, V. and Paci, E. and Pastor, R. W. and Post, C. B. and Pu, J. Z. and Schaefer, M. and Tidor, B. and Venable, R. M. and Woodcock, H. L. and Wu, X. and Yang, W. and York, D. M. and Karplus, M.},
-	date-added = {2024-08-16 10:24:06 -0400},
-	date-modified = {2024-08-16 10:24:09 -0400},
-	doi = {https://doi.org/10.1002/jcc.21287},
+	doi = {10.1002/jcc.21287},
 	journal = {Journal of Computational Chemistry},
 	number = {10},
 	pages = {1545-1614},
 	title = {{CHARMM}: The biomolecular simulation program},
-	url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/jcc.21287},
 	volume = {30},
-	year = {2009},
-	bdsk-url-1 = {https://onlinelibrary.wiley.com/doi/abs/10.1002/jcc.21287},
-	bdsk-url-2 = {https://doi.org/10.1002/jcc.21287}}
+	year = {2009}}
 
 @article{AmberTools,
-	annote = {doi: 10.1021/acs.jcim.3c01153},
[…]
