+++ docs/paper/paper.bib
@@ -9,7 +9,7 @@
 
 
 @article{OpenMM,
-	author = {Eastman, Peter AND Swails, Jason AND Chodera, John D. AND McGibbon, Robert T. AND Zhao, Yutong AND Beauchamp, Kyle A. AND Wang, Lee-Ping AND Simmonett, Andrew C. AND Harrigan, Matthew P. AND Stern, Chaya D. AND Wiewiora, Rafal P. AND Brooks, Bernard R. AND Pande, Vijay S.},
+	author = {Eastman, Peter and Swails, Jason and Chodera, John D. and McGibbon, Robert T. and Zhao, Yutong and Beauchamp, Kyle A. and Wang, Lee-Ping and Simmonett, Andrew C. and Harrigan, Matthew P. and Stern, Chaya D. and Wiewiora, Rafal P. and Brooks, Bernard R. and Pande, Vijay S.},
 	doi = {10.1371/journal.pcbi.1005659},
 	journal = {PLOS Computational Biology},
 	month = {07},
@@ -48,18 +48,14 @@
 
 @article{GROMACS,
 	author = {Mark James Abraham and Teemu Murtola and Roland Schulz and Szil{\'a}rd P{\'a}ll and Jeremy C. Smith and Berk Hess and Erik Lindahl},
-	date-added = {2024-08-16 10:21:09 -0400},
-	date-modified = {2024-08-16 10:21:20 -0400},
-	doi = {https://doi.org/10.1016/j.softx.2015.06.001},
+	doi = {10.1016/j.softx.2015.06.001},
 	issn = {2352-7110},
 	journal = {SoftwareX},
 	pages = {19-25},
 	title = {{GROMACS}: High performance molecular simulations through multi-level parallelism from laptops to supercomputers},
 	url = {https://www.sciencedirect.com/science/article/pii/S2352711015000059},
 	volume = {1-2},
-	year = {2015},
-	bdsk-url-1 = {https://www.sciencedirect.com/science/article/pii/S2352711015000059},
-	bdsk-url-2 = {https://doi.org/10.1016/j.softx.2015.06.001}}
+	year = {2015}}
 
 @article{LAMMPS,
 	author = {A. P. Thompson and H. M. Aktulga and R. Berger and D. S. Bolintineanu and W. M. Brown and P. S. Crozier and P. J. in 't Veld and A. Kohlmeyer and S. G. Moore and T. D. Nguyen and R. Shan and M. J. Stevens and J. Tranchida and C. Trott and S. J. Plimpton},
@@ -193,7 +189,7 @@
 	bdsk-url-1 = {https://doi.org/10.1021/ja01299a050}}
 
 @article{Rappe1991,
-	author = {Rappe, Anthony K and Goddard, William A {III}},
+	author = {Rapp{\'e}, Anthony K and Goddard, William A {III}},
 	doi = {10.1021/j100161a070},
 	journal = {The Journal of Physical Chemistry},
 	month = {apr},
