### 2024-08-16 17b78d5
Reverting bib and updating inline citations
