+++ docs/paper/paper.bib
@@ -1,424 +1,386 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% https://bibdesk.sourceforge.io/
 
-%% Created for Dan Gezelter at 2024-08-16 10:26:30 -0400 
+%% Created for Dan Gezelter at 2024-08-16 10:26:30 -0400
 
 
-%% Saved with string encoding Unicode (UTF-8) 
-
-
-
-@article{OpenMM,
-	author = {Eastman, Peter and Swails, Jason and Chodera, John D. and McGibbon, Robert T. and Zhao, Yutong and Beauchamp, Kyle A. and Wang, Lee-Ping and Simmonett, Andrew C. and Harrigan, Matthew P. and Stern, Chaya D. and Wiewiora, Rafal P. and Brooks, Bernard R. and Pande, Vijay S.},
-	doi = {10.1371/journal.pcbi.1005659},
-	journal = {PLOS Computational Biology},
-	month = {07},
-	number = {7},
-	pages = {1-17},
-	publisher = {Public Library of Science},
-	title = {{OpenMM} 7: Rapid development of high performance algorithms for molecular dynamics},
-	volume = {13},
-	year = {2017}}
-
-@article{CHARMM,
-	author = {Brooks, B. R. and Brooks III, C. L. and Mackerell Jr., A. D. and Nilsson, L. and Petrella, R. J. and Roux, B. and Won, Y. and Archontis, G. and Bartels, C. and Boresch, S. and Caflisch, A. and Caves, L. and Cui, Q. and Dinner, A. R. and Feig, M. and Fischer, S. and Gao, J. and Hodoscek, M. and Im, W. and Kuczera, K. and Lazaridis, T. and Ma, J. and Ovchinnikov, V. and Paci, E. and Pastor, R. W. and Post, C. B. and Pu, J. Z. and Schaefer, M. and Tidor, B. and Venable, R. M. and Woodcock, H. L. and Wu, X. and Yang, W. and York, D. M. and Karplus, M.},
-	doi = {10.1002/jcc.21287},
-	journal = {Journal of Computational Chemistry},
-	number = {10},
-	pages = {1545-1614},
-	title = {{CHARMM}: The biomolecular simulation program},
-	volume = {30},
-	year = {2009}}
+%% Saved with string encoding Unicode (UTF-8)
 
 @article{AmberTools,
[…]
