### 2024-08-16 27b7d8d
Putting missing sentences back.
