### 2024-08-16 #7004 @gezelter
Hi @samwaseda , we've added sentences to the introduction of the paper which (we think) helps to put OpenMD in context.  This has also added references to LAMMPS, CHARMM, Amber, GROMACS, and OpenMM:

"There are a number of high quality molecular dynamics engines that specialize in materials [@LAMMPS] or biomolecular [@CHARMM; @AmberTools] simulations or are models of computational efficiency [@GROMACS; @OpenMM]. In this paper, we provide background on an open source molecular dynamics engine, `OpenMD`, which specializes in complex systems and interfaces and was just released into version 3.1."
