### 2024-08-17 e1cebf5
Cleaning up some samples and code snippets
