### 2024-08-20 94c09f6
Reordering logic in the updateLambda function
