### 2024-08-21 #7004 @crdrisko
Hello @samwaseda, we've rewritten a bit from the beginning of the Statement of Need section to reflect that the rest of the paper highlights some of the different ways in which our engine is unique (compared to the engines mentioned before), namely RNEMD, LangevinHull, DR-EAM, and DSF.

> `OpenMD` builds on the foundations of the Object-Oriented Parallel Simulation Engine (`OOPSE`) program [Meineke2005], rewritten in modern C++. It differs from other contemporary molecular dynamics engines in its focus on complex interfaces. A number of features unique to `OpenMD` facilitate the study of this problem space, including efficient non-equilibrium algorithms, non-periodic simulations, metal polarizability models, and real space electrostatics. The first such feature is Reverse Non-Equilibrium Molecular Dynamics (RNEMD), ...

Hopefully this is what you were looking for, if not, could we get some clarification of what's needed in this section?
