### 2024-08-24 OpenMD/OpenMD#27 @crdrisko
Samples should be fully documented
Some of the samples in our sample directory are missing or have incomplete documentation.

The following samples should have completed `README.md` files in the style of the sample [README](https://github.com/OpenMD/OpenMD/blob/main/samples/README.md-template) template:

- [X] [air](https://github.com/OpenMD/OpenMD/tree/main/samples/air)
- [X] ~~alkane~~
- [X] [aqueousIons](https://github.com/OpenMD/OpenMD/tree/main/samples/aqueousIons)
- [X] [argon](https://github.com/OpenMD/OpenMD/tree/main/samples/argon)
- [X] [bond-order](https://github.com/OpenMD/OpenMD/tree/main/samples/bond-order)
- [X] [builders](https://github.com/OpenMD/OpenMD/tree/main/samples/builders)
- [X] [DR-EAM](https://github.com/OpenMD/OpenMD/tree/main/samples/DR-EAM)
- [X] ~~gbljTest~~
- [X] [graphene](https://github.com/OpenMD/OpenMD/tree/main/samples/graphene)
- [X] [LangevinHull](https://github.com/OpenMD/OpenMD/tree/main/samples/LangevinHull)
- [X] [liquid](https://github.com/OpenMD/OpenMD/tree/main/samples/liquid)
- [X] [Madelung](https://github.com/OpenMD/OpenMD/tree/main/samplesMadelung)
- [x] [metals](https://github.com/OpenMD/OpenMD/tree/main/samples/metals)
- [x] [minimizer](https://github.com/OpenMD/OpenMD/tree/main/samples/minimizer)
- [X] [nanoparticles](https://github.com/OpenMD/OpenMD/tree/main/samples/nanoparticles)
- [X] [protein](https://github.com/OpenMD/OpenMD/tree/main/samples/protein)
- [X] [RNEMD](https://github.com/OpenMD/OpenMD/tree/main/samples/RNEMD/)
- [X] ~~thermoIntegration~~
- [x] [water](https://github.com/OpenMD/OpenMD/tree/main/samples/water)
- [X] [zcons](https://github.com/OpenMD/OpenMD/tree/main/samples/zcons)
- [X] [zeolite](https://github.com/OpenMD/OpenMD/tree/main/samples/zeolite)

N/A

Many samples have been added already using this format, and they can be used as a guide for the above list.

Figures demonstrating the end result of a given sample are encouraged and can be placed in the appropriate directory, [here](https://github.com/OpenMD/OpenMD/tree/main/samples/figures). *Note*: this is generally recommended over an end result file to compare against due to differences in random number generation.
