### 2024-08-24 0dc7347
Removing SPFDynamics since new timestep logic of SPF has made that obsolete
