+++ src/io/Globals.cpp
@@ -248,8 +248,7 @@ namespace OpenMD {
             isEqualIgnoreCase("NPRT") || isEqualIgnoreCase("NPGT") ||
             isEqualIgnoreCase("NGammaT") || isEqualIgnoreCase("NGT") ||
             isEqualIgnoreCase("LANGEVINHULL") || isEqualIgnoreCase("LHULL") ||
-            isEqualIgnoreCase("SMIPD") || isEqualIgnoreCase("LANGEVINPISTON") ||
-            isEqualIgnoreCase("SPF"));
+            isEqualIgnoreCase("SMIPD") || isEqualIgnoreCase("LANGEVINPISTON"));
     CheckParameter(Dt, isPositive());
     CheckParameter(RunTime, isPositive());
     CheckParameter(FinalConfig, isNotEmpty());
