### 2024-08-26 0daea62
Adding some comment stripping and wildCard support for gaff2
