### 2024-08-30 fbb55a4
Adding wildcard matching for inversions
