+++ src/utils/TypeContainer.hpp
@@ -58,6 +58,7 @@
 
 #include "utils/MemoryUtils.hpp"
 #include "utils/Utility.hpp"
+#include "utils/next_combination.hpp"
 
 namespace OpenMD {
 
@@ -136,7 +137,8 @@ namespace OpenMD {
 
       do {
         i = data_.find(permutedKeys);
-        if (i != data_.end()) { return (i->second).second; }
+        if (i != data_.end()) {
+	  return (i->second).second; }
       } while (std::next_permutation(start, permutedKeys.end()));
 
       return NULL;
@@ -176,16 +178,60 @@ namespace OpenMD {
         // replaceWithWildCard can not generate this particular sequence, we
         // have to do it manually
         KeyType allWildCards(SIZE, wildCard);
-        i = data_.find(replacedKey);
+        i = data_.find(allWildCards);
         if (i != data_.end()) { foundTypes.push_back(i->second); }
       }
 
       typename std::vector<ValueType>::iterator j;
       j = std::min_element(foundTypes.begin(), foundTypes.end());
-
       return j == foundTypes.end() ? NULL : j->second;
     }
 
+    ElemPtr permutedFindSkippingFirstElement(KeyType& keys,
+					     const std::string& wildCard) {
+      
+      assert(keys.size() == SIZE);
+      MapTypeIterator i;
+      MapTypeIterator i2;
+      std::vector<KeyTypeIterator> iterCont;
+      std::vector<std::string> replacedKeys;
+      std::vector<ValueType> foundTypes;
+
+      
+      while (replaceWithWildCardSkippingFirstElement(iterCont, keys, replacedKeys, wildCard)) {		
+	KeyType permutedKeys = replacedKeys;
+	// skip the first element:
+	KeyTypeIterator start;
+	start = permutedKeys.begin();
+	++start;
+	std::sort(start, permutedKeys.end());	
+       	do {
+       	  i2 = data_.find(permutedKeys);
+       	  if (i2 != data_.end()) { foundTypes.push_back(i2->second); }
+       	} while (std::next_permutation(start, permutedKeys.end()));
+      }
+            
+      // replaceWithWildCard can not generate this particular sequence, we
+      // have to do it manually
+      
+      KeyType allWildCards(SIZE, wildCard);
+      // but retain first element:
+      allWildCards[0] = keys[0];
+      KeyTypeIterator start;
+      start = allWildCards.begin();
+      ++start;
+      std::sort(start, allWildCards.end());
+      do {
+	i2 = data_.find(allWildCards);
+	if (i2 != data_.end()) { foundTypes.push_back(i2->second); }
+      } while (std::next_permutation(start, allWildCards.end()));
+      
+      
+      typename std::vector<ValueType>::iterator j;
+      j = std::min_element(foundTypes.begin(), foundTypes.end());
+      return j == foundTypes.end() ? NULL : j->second;
+    }
+    
     size_t size() { return data_.size(); }
 
     ElemPtr beginType(MapTypeIterator& i) {
