### 2024-09-12 #7004 @samwaseda
I haven't fully been able to test the software for fields I'm familiar with as there's still some documentation missing in the example folder (v.s.).
