### 2024-09-12 ccc6703
Reworking how SPF-RNEMD functions during a timestep
