### 2024-09-12 725268b
Adding EAM README
