### 2024-09-12 72cbe1c
Finishing up the protein sample's expected results section
