+++ samples/nanoparticles/README.md
@@ -1,145 +1,151 @@
 # Metallic nanoparticles
 
-
 ## Background Information
 
 In these examples, we’ll build gold nanoparticles and equilibrate them to a temperature of 300K. 
 
 ## Instructions
-### Building a spherical metal nanoparticle
+
+### Example 1
+
+Here we'll look at building a spherical metal nanoparticle:
 
 1. We need to start with an input `.omd` file for the metal of your choice. This directory has a `gold.omd` for this example. It looks like:
-```
-<OpenMD>
-  <MetaData>
-molecule{
-  name = "Au";
+    ```
+    <OpenMD>
+      <MetaData>
+    molecule{
+      name = "Au";
 
-  atom[0]{
-    type = "Au";
-    position(0.0, 0.0, 0.0);
-  }
-}
+      atom[0]{
+        type = "Au";
+        position(0.0, 0.0, 0.0);
+      }
+    }
 
-component{
-  type = "Au";
-  nMol = 1;
-}
+    component{
+      type = "Au";
+      nMol = 1;
+    }
 
-forceField = "SC";
-forceFieldFileName = "SuttonChen.QSC.frc";
-  </MetaData>
-</OpenMD>
-```
+    forceField = "SC";
+    forceFieldFileName = "SuttonChen.QSC.frc";
+      </MetaData>
+    </OpenMD>
+    ```
 
 2. Now to build the nanoparticle. We’ll choose a radius of 15 Å and use the lattice constant of gold, 4.08 Å. We’ll call our initial structure `NP15.omd`. 
-```
-nanoparticleBuilder --latticeConstant=4.08 --radius=15 gold.omd -o NP15.omd
-```
+    ```
+    nanoparticleBuilder --latticeConstant=4.08 --radius=15 gold.omd -o NP15.omd
+    ```
 
 3. To take a look at the structure we’ve just created, you can use the following command:
-```
-Dump2XYZ -i NP15.omd
-```
-to create a file called `NP15.xyz`, which can be viewed in VMD, Jmol, or any other chemical structure viewer.
+    ```
+    Dump2XYZ -i NP15.omd
+    ```
+    to create a file called `NP15.xyz`, which can be viewed in VMD, Jmol, or any other chemical structure viewer.
 
 4. Add the following lines to the new NP15.omd file below the `forceFieldFileName` line. This sets some details for the simulation. 
-```
-ensemble = "LHull";
-targetTemp = 5;
-targetPressure = 1;
-viscosity = 0.1;
-dt = 4.0;
-runTime = 2E5;
+    ```
+    ensemble = "LHull";
+    targetTemp = 5;
+    targetPressure = 1;
+    viscosity = 0.1;
+    dt = 4.0;
+    runTime = 2E5;
 
-sampleTime = 2000.0;
-statusTime = 4;
-seed = 985456376;
+    sampleTime = 2000.0;
+    statusTime = 4;
+    seed = 985456376;
 
-usePeriodicBoundaryConditions = "false";
-tauThermostat = 1E3;
-tauBarostat = 5E3;
-```
+    usePeriodicBoundaryConditions = "false";
+    tauThermostat = 1E3;
+    tauBarostat = 5E3;
+    ```
 
 5. NanoparticleBuilder carves a nanoparticle of our chosen radius out of a perfect gold crystal. We need to give the atoms some initial velocities before we start equilibrating. We’ll start it out at 5K:
-```
-thermalizer -t 5 NP15.omd -o NP15_5K.omd
-```
+    ```
+    thermalizer -t 5 NP15.omd -o NP15_5K.omd
+    ```
 
[…]
