+++ samples/protein/README.md
@@ -11,75 +11,107 @@ One of the major tasks facing a new user of OpenMD is getting their molecule of 
 2. Alternatively (and this is the route we’ll take here), you can download a protein structure directly from the Protein Data Bank. We’ve picked a short pentapeptide (the neurotransmitter met-enkephalin) to use as an example. The PDB-ID for this protein is 1PLW. The text version of the PDB file is located in this sample directory.
 
 3. Use the atom2md program to convert the structure into a format that can be read by OpenMD:
-```
-atom2md -ipdb 1PLW.pdb
-```
-This command will create an incomplete OpenMD file called `1PLW.omd` that must be edited before it can be used.
+    ```
+    atom2md -ipdb 1PLW.pdb
+    ```
+    This command will create an incomplete OpenMD file called `1PLW.omd` that must be edited before it can be used.
 
 4. OpenMD can use the Amber force field for protein simulations. If you are using this force field, the N-terminal and C-terminal ends of the protein should be modified to use the correct atom types. This peptide has the sequence “Tyr-Gly-Gly-Phe-Met”, so Tyrosine is the N-terminal end and Methionine is the C-terminal end.
 
-In the `omd` file, the atom typing for the first residue must be changed from TYR-X to NTYR-X while for the last residue, it must be changed from MET-X to CMET-X. The special N- and C-terminal atom types usually only apply to the N, CA, C, O, OXT, HN, and HA base types. Other atom types can usually be left alone.
+    In the `omd` file, the atom typing for the first residue must be changed from TYR-X to NTYR-X while for the last residue, it must be changed from MET-X to CMET-X. The special N- and C-terminal atom types usually only apply to the N, CA, C, O, OXT, HN, and HA base types. Other atom types can usually be left alone.
 
 5. Some PDB files specify more details on the atom name record than the Amber force field can use. In particular, the PHE-CD1 and PHE-CD2 types should be down-specified to PHE-CD. Similar down-specification can be done for PHE-CE1, PHE-CE2, PHE-HD1, PHE-HD2, PHE-HE1, PHE-HE2 atom types.
 
 6. At this point it is also a good idea to change the name of the molecule to something descriptive (perhaps “metenk”). This should be done in two places; once in the molecule description and another time in the component block.
 
 7. Before the simulation can run, add a forceField line after the component block:
-```
-forceField = "Amber";
-```
-At this stage, you should be able to run OpenMD on the file to check to make sure your hand-crafted atom typing can be matched up with types known by the force field:
-```
-openmd 1PLW.md
-```
-Correct any unknown atom types, and repeat until you get an error about the “Integrator Factory”.
+    ```
+    forceField = "Amber";
+    ```
+    At this stage, you should be able to run OpenMD on the file to check to make sure your hand-crafted atom typing can be matched up with types known by the force field:
+    ```
+    openmd 1PLW.md
+    ```
+    Correct any unknown atom types, and repeat until you get an error about the “Integrator Factory”.
 
 8. Add the following lines below the forceField line:
-```
-ensemble = NVT;
-cutoffMethod = "shifted_force";
-electrostaticScreeningMethod = "damped";
-cutoffRadius = 10;
-dampingAlpha = 0.18;
-targetTemp = 300;
-tauThermostat = 1000;
-dt = 1.0;
-runTime = 1e4;
-tempSet = "false";
-sampleTime = 100;
-statusTime = 10;
-```
+    ```
+    ensemble = NVT;
+    cutoffMethod = "shifted_force";
+    electrostaticScreeningMethod = "damped";
+    cutoffRadius = 10;
+    dampingAlpha = 0.18;
[…]
