+++ samples/metals/Sutton-Chen/README.md
@@ -1,0 +1,128 @@
+# Simulations using the Sutton-Chen potential
+ 
+The Sutton-Chen (SC) potential has been used to
+study a wide range of phenomena in metals: 
+
+> A. P. Sutton and J. Chen. Long-range Finnis Sinclair potentials. *Phil. Mag. Lett.*, **61**:139–146, 1990.
+ 
+Although it has the same basic form as the EAM potential, 
+the Sutton-Chen model requires a simpler set of parameters,
+
+$$
+U_{tot}=\sum _{i}\left[ \frac{1}{2}\sum _{j\neq
+i}\epsilon_{ij}V^{pair}_{ij}(r_{ij})-c_{i}\epsilon_{ii}\sqrt{\rho_{i}}\right] ,
+$$
+
+where $V^{pair}_{ij}$ and $\rho_{i}$ are given by 
+$$
+V^{pair}_{ij}(r)=\left(
+\frac{\alpha_{ij}}{r_{ij}}\right)^{n_{ij}} \hspace{1in} \rho_{i}=\sum_{j\neq i}\left(
+\frac{\alpha_{ij}}{r_{ij}}\right) ^{m_{ij}}
+$$
+
+$V^{pair}_{ij}$ is a repulsive pairwise potential that accounts for
+interactions of the pseudo-atom cores.  The $\sqrt{\rho_i}$ term in
+the first equation is an attractive many-body potential that models
+the interactions between the valence electrons and the cores of the
+pseudo-atoms.  $\epsilon_{ij}$, $\epsilon_{ii}$, $c_i$ and
+$\alpha_{ij}$ are parameters used to tune the potential for different
+transition metals.
+
+The SC potential form has also been parameterized by Qi, *et
+al.*:
+
+> Y. Qi, T. &Ccedil;aǧin, Y. Kimura, and W. A. Goddard III. Molecular dynamics simulations of glass formation and crystallization in binary liquid metals: Cu-Ag and Cu-Ni. *Phys. Rev. B*, **59**(5):3527–3533, 1999.
+
+These parameters were obtained via empirical and *ab initio* 
+calculations to match structural features of the FCC crystal.  This variant is called the `QSC` variant in `OpenMD` 
+
+## Background Information
+
+The two sample files in this directory describe identical short (3 ps) simulations of bulk gold in a FCC crystal.  The only differences are in the force field variants, but in this example, we are using the file name of the force field to select the variant:
+
+```
+forceFieldFileName = "SuttonChen.QSC.frc";
+```
+for the QSC sample and 
+```
+forceFieldFileName = "SuttonChen.frc";
+```
+for the original Sutton-Chen parameters. These samples also illustrate another feature, periodic thermalization of the atomic velocitites. The initial atomic velocities in these samples are all 0, so to provide kinetic energy, we add these lines to the `<MetaData>` section:
+
+```
+targetTemp = 400.0;
+thermalTime = 400;
+seed = 985456376;
+tempSet="true";
+```
+These four parameters, set a target temperature of 400K, tell `OpenMD` to resample velocities (and to do so every 400 fs) and to use a random seed so that the process is reproducible.
+
+## Instructions
+
+To run the samples, we would do the following commands:
+```bash
+openmd Au_bulk_SC.omd   # simulate using original Sutton-Chen parameters
+openmd Au_bulk_QSC.omd  # simulate using the Quantum Sutton-Chen parameters
+```
+
+Another utility built in to `OpenMD` can be used to look at the elastic constants of gold:
+
+```
+elasticConstants -i Au_bulk_QSC.omd -b --method="energy"
+```
+
+## Expected Output
+
[…]
