### 2024-09-12 0801520
Equation issues
