+++ samples/metals/Sutton-Chen/README.md
@@ -9,13 +9,13 @@ Although it has the same basic form as the EAM potential,
 the Sutton-Chen model requires a simpler set of parameters,
 
 $$
-U_{tot}=\sum_{i}\left[\frac{1}{2}\sum_{j\neq i}\epsilon_{ij}V^{pair}_{ij}(r_{ij})-c_{i}\epsilon_{ii}\sqrt{\rho_{i}} \right] ,
+U_\text{tot}=\sum_{i}\left[ \frac{1}{2} \sum_{j\neq i} \epsilon_{ij} V^\text{pair}_{ij}(r_{ij}) - c_{i} \epsilon_{ii} \sqrt{\rho_{i}} \right]
 $$
 
-where $V^{pair}_{ij}$ and $\rho_{i}$ are given by 
+where $V^\text{pair}_{ij}$ and $\rho_{i}$ are given by 
 $$
-V^{pair}_{ij}(r)=\left(
-\frac{\alpha_{ij}}{r_{ij}}\right)^{n_{ij}} \hspace{1in} \rho_{i}=\sum_{j\neq i}\left(
+V^\text{pair}_{ij}(r)=\left(
+\frac{\alpha_{ij}}{r_{ij}}\right)^{n_{ij}} ~~~~~~~~~~~~~ \rho_{i}=\sum_{j\neq i}\left(
 \frac{\alpha_{ij}}{r_{ij}}\right) ^{m_{ij}}
 $$
 
