### 2024-09-13 OpenMD/OpenMD#27 @crdrisko
We've removed the alkane, gbljTest, and thermoIntegration sample directories as they dealt with older, less interesting samples. These were either better covered in other places or not useful enough to keep.
