+++ samples/alkane/README.md
@@ -1,12 +1,0 @@
-# Alkane
-
-## Background Information
-
-OpenMD supports the Dipolar Unified-atom Force Field (DUFF) with many
-of the parameters coming from the
-[TraPPE-UA](http://chem-siepmann.oit.umn.edu/siepmann/trappe/index.html)
-forcefield from Siepmann's group.
-
-## Instructions
-
-## Expected Output
