+++ samples/nanoparticles/README.md
@@ -2,7 +2,7 @@
 
 ## Background Information
 
-In these examples, we’ll build gold nanoparticles and equilibrate them to a temperature of 300K. 
+In these examples, we’ll build gold nanoparticles and equilibrate them to a temperature of 300K.
 
 ## Instructions
 
@@ -34,7 +34,7 @@ Here we'll look at building a spherical metal nanoparticle:
     </OpenMD>
     ```
 
-2. Now to build the nanoparticle. We’ll choose a radius of 15 Å and use the lattice constant of gold, 4.08 Å. We’ll call our initial structure `NP15.omd`. 
+2. Now to build the nanoparticle. We’ll choose a radius of 15 Å and use the lattice constant of gold, 4.08 Å. We’ll call our initial structure `NP15.omd`.
     ```
     nanoparticleBuilder --latticeConstant=4.08 --radius=15 gold.omd -o NP15.omd
     ```
@@ -45,7 +45,7 @@ Here we'll look at building a spherical metal nanoparticle:
     ```
     to create a file called `NP15.xyz`, which can be viewed in VMD, Jmol, or any other chemical structure viewer.
 
-4. Add the following lines to the new NP15.omd file below the `forceFieldFileName` line. This sets some details for the simulation. 
+4. Add the following lines to the new NP15.omd file below the `forceFieldFileName` line. This sets some details for the simulation.
     ```
     ensemble = "LHull";
     targetTemp = 5;
@@ -84,7 +84,7 @@ Here we'll look at building a spherical metal nanoparticle:
     ```
 
 9. To continue with the equilibration we need to change the `targetTemp` of `NP15_100K.omd`. We’ll increase it to 100 and run the `NP15_100K.omd` file.
-    
+
 10. We’ll continue the procedure of copying the .eor file to a new .omd file and increasing the temperature until we’ve reached 300K. Temperature increases of 50 – 100K and simulation times of 100 – 200 ps are reasonable.
 
 ### Example 2
@@ -95,17 +95,17 @@ In this example, we'll be building an icosahedral metal nanoparticle:
 
 2. To create the particle coordinates, use the command:
     ```
-    icosahedralBuilder -o file.omd --shell=8 --latticeConstant=4.08 --ico gold.omd
+    icosahedralBuilder -o ICO8.omd --shell=8 --latticeConstant=4.08 --ico gold.omd
     ```
-    where `file.omd` is the name of your icosahedra, 8 is the number of shells (in this example), 4.08 is the lattice constant in Angstroms for the metal (4.08 is the correct value for gold), and the skeletal OpenMD file above is called `gold.omd`
+    where `ICO8.omd` is the name of your icosahedra, 8 is the number of shells (in this example), 4.08 is the lattice constant in Angstroms for the metal (4.08 is the correct value for gold), and the skeletal OpenMD file above is called `gold.omd`
 
 3. To make sure this icosahedra is the right size for your purposes; use Dump2XYZ to create an xyz file:
     ```
-    Dump2XYZ -i file.omd
+    Dump2XYZ -i ICO8.omd
     ```
     and view the file in VMD or Jmol (where you might want to measure the diameter of the particle).
 
-4. To run the icosahedra in a Langevin Hull and heat the system, we’ll start by inserting the following excerpt after the `forceFieldFileName` in `file.omd`.
+4. To run the icosahedra in a Langevin Hull and heat the system, we’ll start by inserting the following excerpt after the `forceFieldFileName` in `ICO8.omd`.
     ```
     ensemble = "LHull";
     targetTemp = 5;
@@ -123,24 +123,24 @@ In this example, we'll be building an icosahedral metal nanoparticle:
 
 5. Before we can equilibrate the system, the atoms need initial velocities, which we can set using the thermalizer command.
     ```
[…]
