### 2024-09-16 bcb11fa
Failed kick counter in RNEMD should go up if we can't find a particle to place
