### 2024-09-17 fae4ace
Finishing up EAM nanoparticle samples
