### 2024-09-17 73c7c5e
Changing up analysis section of the liquid sample
