### 2024-09-18 507c974
Install scripts in samples directory in the correct locations
