+++ samples/metals/surfaces/README.md
@@ -2,22 +2,221 @@
 
 ## Background Information
 
-Contained here are examples of (111) cut surfaces of the coinage
-metals, as well as a few of the catalytically active metals.
+One common task done with OpenMD is the simulation of metallic 
+surfaces. Here, we show how to create and simulate particular facets
+of FCC metals, sometimes in contact with liquids or other molecules.
 
 ### slabBuilder
 
-OpenMD also has a utility script which makes creation of these types
-of systems trivial. `slabBuilder` is a python script which generates
-*sc*, *bcc*, and *fcc* lattices and cleaves the crystals along a
+OpenMD has a utility script which makes creation of these types
+of systems relatively easy. `slabBuilder` is a python script which 
+generates *sc*, *bcc*, and *fcc* lattices and cleaves the crystals along a
 desired *(hkl)* plane. The systems are then reoriented such that the
 cleaved facet is presented to the z-dimension of the simulation
 box. `slabBuilder` comes with a help message when passed `-h` or
 `--help`.
 
-## Instructions
+### waterBoxer and omd-solvator
+
+Another helpful utility called `waterBoxer` creates initial coordinates 
+of water boxes.  In what follows we're going to create a metal interface
+and a matching box of water, and then we'll combine them together using
+`omd-solvator`
+
+## Instructions for Au(111) - Water Interface
+
+Let's first build a gold slab with the 111 facet facing the 
+*z*-direction, and with some vaccuum space above and below the slab.
+Gold forms fcc lattices, and the lattice constant is 4.08 
+angstroms. We'll specify the facet with the `--hkl` flag, and how many 
+unit cell repeats we want with the `-r` flag.  This command creates an 
+output `omd` file and also puts the coordinates in a matching `xyz` file:
+
+```
+slabBuilder -l fcc -c 4.08 -o gold111.omd --hkl 1 1 1 -r 3 4 3 -e Au -v true -x gold111.xyz
+```
+The `slabBuilder` command prints out the dimensions of the slab it created, but the size of the periodic box is also given on the `Hmat` line in the generated `gold111.omd` file:
+```
+     Hmat: {{ 29.981754, 0.000000, 0.000000 }, { 0.000000, 23.079965, 0.000000 }, { 0.000000, 0.000000, 63.600906 }}
+```
+
+Next, we want to generate a box of water that is exactly the same size as the gold slab:
+```
+waterBoxer -d 1.0 -w SPCE -x 29.981754 -y 23.079965 -z 63.600906 -r -m
+Dump2XYZ -i freshWater.omd -b 
+```
+This creates a file called `freshWater.omd` which has dimensions that 
+exactly match the gold slab, and the second command allows us to 
+visualize these coordinates in the `freshWater.xyz` using Jmol or VMD. 
+Here we've randomized the orientations of the water molecules (`-r`) and 
+have printed a `water.inc` include file (`-m`)
+
+The next step is to combine the two files together to create a metal - water interface in a `combined.omd` file.
+
+```
+omd-solvator --solute=gold111.omd --solvent=freshWater.omd -r 3.5 -o combined.omd -n 1 -p 1
+```
+We can't run this file yet because force field and molecular definitions need to be added.  Edit the `combined.omd` file so that the `<MetaData>` section near the top looks like this:
+```C++
+   <MetaData>
+#include 'metals.inc'  // includes the molecule definitions for metals
+#include 'water.inc'   // includes the molecule definitions for water models
+
+component{
[…]
