### 2024-09-20 92573ae
Fixing base type checking for fictitious sites in ReplacementVisitor
