### 2024-09-23 cc41e87
Adding the minimizer sample documentation
