+++ samples/minimizer/tip4p_two.in
@@ -1,4 +1,0 @@
-2
-10000;	60	0	0;	0	60	0;	0	0	60;	0.136949	1115.21;
-TIP4P_RB_0	-2.2274e+00 -2.5720e+00 -2.8204e+00	0.000000	0.000000	-0.00000	-1.1951e-01  8.5679e-01  2.2958e-01  4.4601e-01		0.000022	-0.000027	0.000007
-TIP4P_RB_0	-2.2274e+00 -2.5720e+00  0.0000e+00	-0.000000	-0.000000	0.000000	  3.9989e-01  7.6817e-01 -4.4351e-01 -2.3087e-01	-0.000045	-0.000007	-0.000007
