### 2024-09-24 e8c14a7
Added dimer sample README.md, started work on ice_surfaces sample
