### 2024-09-24 c172721
Adding confidence intervals to tetrahedrality by z.
