### 2024-09-24 3db3cf3
Fixed NaN issue in confidence interval
