+++ samples/water/liquid/README.md
@@ -1,0 +1,153 @@
+# Liquid water simulations with two water models
+
+This directory contains equilibrated configurations for simulating
+liquid water with five different water models, SPC/E and TIP4P, which 
+are multi-site models, and SSD, SSD/E, and SSD/RF which are single-site 
+models. These simulations are relatively small (256 molecules) and in 
+periodic box that is 19.7 angstroms on a side. We'll carry out a 
+relatively short (1e5 fs = 100ps) simulation and concentrate on how 
+to extract interesting data from the trajectories. One interesting observation
+is that the SSD-based simulations run roughly 9-16 times faster than the
+multi-site simulations.
+
+The SSD/RF model was tuned for use with a reaction field for electrostatic 
+interactions, so you might note that the `ssdrf.omd` file has these 
+parameters:
+```
+electrostaticSummationMethod = "reaction_field";
+electrostaticScreeningMethod = "undamped";
+dielectric = 80.0;
+cutoffRadius = 9.0;
+```
+while the other simulations use:
+```
+cutoffMethod = "shifted_force";
+electrostaticScreeningMethod = "damped";
+dampingAlpha = 0.18;
+cutoffRadius = 9.0;
+```
+
+Because we're going to extract some time correlation function information,
+we're doing these simulations in the NVE ensemble, and we're going to sample
+configurations frequently enough to generate some useful time correlation
+functions.  Here are the relevant keywords in the `<MetaData>` section:
+```
+ensemble = NVE;
+runTime = 1e5;
+sampleTime = 100;
+statusTime = 10;
+```
+## The water models
+
+The following equilibrated liquid configurations are provided,
+
++ `spce.omd` - SPC/E 3-site water molecules
++ `tp4.omd` - TIP4P 4-site water molecules
++ `ssd.omd` - SSD single site water molecules
++ `ssde.omd` - SSD/E single site water molecules
++ `ssdrf.omd` - SSD/RF single site water molecules
+
+For more information on the water models, we suggest:
+
+> The missing term in effective pair potentials H. J. C. Berendsen, J. R. Grigera, and T. P. Straatsma, The Journal of Physical Chemistry 1987 91 (24), 6269-6271, DOI: [10.1021/j100308a038](https://doi.org/10.1021/j100308a038) |
+
+and 
+
+> William L. Jorgensen, Jayaraman Chandrasekhar, Jeffry D. Madura, Roger W. Impey, Michael L. Klein; Comparison of simple potential functions for simulating liquid water. J. Chem. Phys. 15 July 1983; 79 (2): 926–935. DOI: [10.1063/1.445869](https://doi.org/10.1063/1.445869) 
+
+and 
+
+> Ming-Liang Tan, Justin T. Fischer, Amalendu Chandra, Bernard R. Brooks, Toshiko Ichiye,
+A temperature of maximum density in soft sticky dipole water, Chemical Physics Letters,
+376(5–6), pp. 646-652 (2003). DOI: [10.1016/S0009-2614(03)01044-3](https://doi.org/10.1016/S0009-2614(03)01044-3) 
+
+and
+
+> Christopher J. Fennell, J. Daniel Gezelter; On the structural and transport properties of the soft sticky dipole and related single-point water models. J. Chem. Phys. 15 May 2004; 120 (19): 9175–9184. DOI: [10.1063/1.1697381](https://doi.org/10.1063/1.1697381) 
+
+## Instructions
+
+These samples will be run with multiple processors:
+```
+mpirun -np 4 openmd_MPI spce.omd
+mpirun -np 4 openmd_MPI tp4.omd
[…]
