### 2024-09-25 d7e5aa7
Removing old ssd sample
