### 2024-09-27 846b626
Adding ice_surfaces sample
