+++ samples/water/ice_surfaces/README.md
@@ -1,0 +1,153 @@
+# Structures for simulating Ice-Ih interfaces with liquid water
+
+This directory contains four prepared systems with different ice facets 
+exposed to liquid water. They were prepared using a block of Ice-Ih that
+has been cleaved along one particular facet. These blocks were then  
+rotated so that the prepared facet was oriented to the + and - *z* 
+direction of the periodic box, with a vaccuum layer separating periodic 
+replicas.  For instructions on preparing these blocks, see the 
+[`README.md`](../iceCrystals/README.md) file in the 
+[iceCrystals](../iceCrystals) directory.
+
+A second box of liquid water using the same box geometry and water 
+model was also prepared, and then the two boxes were combined, eliminating
+the liquid phase water molecules wherever they overlapped with ice-phase
+molecules.
+
+The systems were equilibrated to a temperature just below the melting
+point of the water model. We are using the `TIP4P/Ice` water model which can
+be found in this paper:
+
+> J. L. F. Abascal, E. Sanz, R. García Fernández, C. Vega; A potential model for the study of ices and amorphous water: TIP4P/Ice. J. Chem. Phys. 15 June 2005; 122 (23): 234511. [10.1063/1.1931662](https://doi.org/10.1063/1.1931662)
+
+Equilibration was done with a weak thermal flux that was used to
+transfer kinetic energy from the ice region into the liquid, keeping the
+ice at a slightly lower temperature than the liquid. For the 
+simulations in these samples, that flux has been turned off, but 
+you can see what was used in previous steps in the `omd` files:
+
+```
+RNEMD{
+  useRNEMD = false;
+  objectSelection = "select TIP4P-Ice_RB_0";
+  exchangeTime = 2;
+  fluxType = "KE";
+  method = "VSS";
+  outputBins = 100;
+  kineticFlux = -2.0e-6;
+  outputFields = "Z|TEMPERATURE|VELOCITY|DENSITY";
+}
+```
+If you want to experiment with reverse non-equilibrium molecular
+dynamics (RNEMD), you can change the block above to include
+`useRNEMD = true;` which will re-enable the kinetic energy flux.
+
+## Facets of ice
+
+The following solvated ice-Ih facets were prepared:
+
++ Basal {0001} facet - `solvatedBasal.omd`
++ Prismatic {001̅0} facet - `sovatedPrism.omd`
++ Secondary Prism {112̅0} facet - `sovatedSecPrism.omd`
++ Pyramidal {202̅1} facet - `solvatedPyramidal.omd`
+
+Here's a side view of the solvate basal facet generated with:
+```
+Dump2XYZ -i solvatedBasal.omd -b -m
+jmol solvatedBasal.xyz
+```
+<img src="../../figures/solvateBasal.png" alt="image" width="500" height="auto">
+
+
+## Instructions 
+
+We're going to run one 100 ps simulation of a solvated basal facet and then we'll 
+analyze the local tetrahedral order parameter during the simulation. **Note
+that these simulations are quite large and take a significant amount of processor
+time.**
+
+1. First let's run that simulation. If you have more than 4 processors available, now
+would be a good time to use them:
+
+```
+mpirun -np 4 openmd_MPI solvatedBasal.omd
+```
+
+2. If all went well, you should get a report showing that the average temperature stayed 
+at 267 K, just below the melting point:
+```
+###############################################################################
+# Status Report:                                                              #
+#              Total Time:      100000 fs                                     #
+#       Number of Samples:        1001                                        #
[…]
