### 2024-09-30 70d7055
Missing figure in fieldScreening sample has been added.
