### 2024-10-01 ebe1fcd
Cleaning up install process
