### 2024-10-03 #7004 @prgupta-cpfam
@gezelter, can you please direct me to the automated tests for REMD and Real space electrostatics implementations?
