### 2024-10-03 999c7c8
Adding consistent syntax highlighting to markdown files and reformatting some samples
