+++ samples/nanoparticles/README.md
@@ -11,7 +11,7 @@ In these examples, we’ll build gold nanoparticles and equilibrate them to a te
 Here we'll look at building a spherical metal nanoparticle:
 
 1. We need to start with an input `.omd` file for the metal of your choice. This directory has a `gold.omd` for this example. It looks like:
-    ```
+    ```C++
     <OpenMD>
       <MetaData>
     molecule{
@@ -35,18 +35,18 @@ Here we'll look at building a spherical metal nanoparticle:
     ```
 
 2. Now to build the nanoparticle. We’ll choose a radius of 15 Å and use the lattice constant of gold, 4.08 Å. We’ll call our initial structure `NP15.omd`.
-    ```
+    ```bash
     nanoparticleBuilder --latticeConstant=4.08 --radius=15 gold.omd -o NP15.omd
     ```
 
 3. To take a look at the structure we’ve just created, you can use the following command:
-    ```
+    ```bash
     Dump2XYZ -i NP15.omd
     ```
     to create a file called `NP15.xyz`, which can be viewed in VMD, Jmol, or any other chemical structure viewer.
 
 4. Add the following lines to the new NP15.omd file below the `forceFieldFileName` line. This sets some details for the simulation.
-    ```
+    ```C++
     ensemble = "LHull";
     targetTemp = 5;
     targetPressure = 1;
@@ -64,22 +64,22 @@ Here we'll look at building a spherical metal nanoparticle:
     ```
 
 5. NanoparticleBuilder carves a nanoparticle of our chosen radius out of a perfect gold crystal. We need to give the atoms some initial velocities before we start equilibrating. We’ll start it out at 5K:
-    ```
+    ```bash
     thermalizer -t 5 NP15.omd -o NP15_5K.omd
     ```
 
 6. For the first step in the equilibration we need to let the gold lattice structurally relax. `NP15_5K.omd` can now be run:
-    ```
+    ```bash
     openmd NP15_5K.omd
     ```
 
 7. Running the simulation will create several new files. `NP15_5K.dump` contains the trajectory of the simulation. Statistics such as temperature, pressure, and energy will be recorded in the `NP15_5K.stat` file and can be viewed using:
-    ```
+    ```bash
     xmgrace -nxy NP15_5K.stat
     ```
 
 8. The end-of-run file `NP15_5K.eor` stores the last configuration of the simulation. We’ll copy it to a new .omd file.
-    ```
+    ```bash
     cp NP15_5K.eor NP15_100K.omd
     ```
 
@@ -94,19 +94,19 @@ In this example, we'll be building an icosahedral metal nanoparticle:
 1. To start the process, a metal file (`gold.omd`) is needed to describe the material composition of the particle.  We'll use the same file from the spherical example above.
 
 2. To create the particle coordinates, use the command:
-    ```
+    ```bash
     icosahedralBuilder -o ICO8.omd --shell=8 --latticeConstant=4.08 --ico gold.omd
     ```
     where `ICO8.omd` is the name of your icosahedra, 8 is the number of shells (in this example), 4.08 is the lattice constant in Angstroms for the metal (4.08 is the correct value for gold), and the skeletal OpenMD file above is called `gold.omd`
 
 3. To make sure this icosahedra is the right size for your purposes; use Dump2XYZ to create an xyz file:
-    ```
+    ```bash
     Dump2XYZ -i ICO8.omd
     ```
     and view the file in VMD or Jmol (where you might want to measure the diameter of the particle).
 
 4. To run the icosahedra in a Langevin Hull and heat the system, we’ll start by inserting the following excerpt after the `forceFieldFileName` in `ICO8.omd`.
-    ```
+    ```C++
     ensemble = "LHull";
     targetTemp = 5;
     targetPressure = 1;
[…]
