### 2024-10-05 29409ec
Fixed segmentation fault caused by uninitialized variable and removed explicit info->update calls, this closes #24
