+++ src/optimization/ObjectiveFunction.hpp
@@ -1,7 +1,7 @@
 /* -*- mode: c++; tab-width: 4; indent-tabs-mode: nil; c-basic-offset: 4 -*- */
 
 /*
-  Copyright (C) 2001, 2002, 2003 Nicolas Di C�sar�
+  Copyright (C) 2001, 2002, 2003 Nicolas Di Cesare
 
   This file is part of QuantLib, a free-software/open-source library
   for financial quantitative analysts and developers - http://quantlib.org/
