+++ samples/metals/Sutton-Chen/README.md
@@ -97,9 +97,9 @@ Original Box Geometry (Angstroms):
 
 Optimized Box Geometry (Angstroms):
 
-      28.4179 -1.37164e-18 -2.98806e-19 
- -1.37164e-18      28.4179 -2.10653e-18 
- -2.98806e-19 -2.10653e-18      28.4179 
+      28.4179 -6.01519e-18 -2.91197e-18 
+ -6.01519e-18      28.4179 -1.17687e-18 
+ -2.91197e-18 -1.17687e-18      28.4179 
 
 Elastic Tensor (GPa):
 
