### 2024-10-09 #7004 @prgupta-cpfam
Hi @srmnitc . I have gone through the Madelung directory as suggested by @gezelter. I was actually waiting for the details on RNEMD as asked by @gezelter.
