### 2024-10-09 3847213
First part of the RNEMD sample changes
