+++ samples/RNEMD/README.md
@@ -33,75 +33,124 @@ shorter simulation times to obtain converged results for transport properties.
 
 ## Instructions
 
-Unless explicitly mentioned, you can run `openmd` on all these examples
-mentioned below with the following command:
-
-```bash
-mpirun -np 4 openmd_MPI <sample_file>.omd
-```
-
 ### Example 1
 
-The file `shearWater.omd` is a box of 1500 SPC/E water molecules
-which has the momentum flux functionality of RNEMD turned on. Notice
-in the RNEMD block of the (.omd) file that
+For our first test of the functionality in the RNEMD module, we'll start 
+with a simple test of applying a kinetic energy flux on a system of 2744 
+Argon atoms. Take a look at the RNEMD block in the `2744.omd` file:
 
 ```C++
 RNEMD {
-  fluxType = "Px";
-  momentumFlux = 6.0e-7;
-  // ...
+  useRNEMD = true;
+  exchangeTime = 1.0;
+  outputBins = 40;
+  objectSelection = "select Ar";
+  fluxType = "KE";
+  method = "VSS";
+  kineticFlux = -2e-9;
+  outputFields = "Z|TEMPERATURE|VELOCITY|DENSITY";
 }
 ```
 
-With these parameters, the applied flux will be an x-axis momentum
-flux transferred across the z-dimension of the box. Using this
-functionality, one is able to measure the *shear viscosity* of the
-liquid at the simulated temperature by relating the imposed momentum
-flux to the system's gradient response of the velocity.
+A couple things of note here are the `fluxType` and `method` parameters. 
+These signify that the algorithm type we're using is the VSS-RNEMD algorithm 
+which has been shown to be quite good in measuring thermal conductivities 
+from molecular dynamics simulations, hence the `KE`, a.k.a. thermal, flux type.
+
+To run this simulation, we'd reccommend adding a few extra cores with the 
+following command:
+
+```bash
+mpirun -np 4 openmd_MPI 2744.omd
+```
 
 ### Example 2
 
-It is also possible to measure the *thermal conductivity* of a
-material using the RNEMD functionality in OpenMD. As an example,
-`graphene.omd` is an (.omd) file where two sheets of graphene have a
-thermal flux applied across the long axis of the sheet. In the (.omd)
-file the fluxType has been set to a kinetic energy flux, and also that
-the kineticFlux is defined.
+The file `2744_shear.omd` is a box of 2744 Argon atoms which has a
+simultaneous momentum and kinetic energy flux through the box. In the
+middle of the simulation cell is a lattice structure which has interesting
+implications for the velocity profiles. Notice this in the RNEMD block of 
+the (`.omd`) file that we've swapped our `fluxType` to these combined thermal 
+and momentum fluxes:
 
 ```C++
 RNEMD {
-  fluxType = "KE";
-  kineticFlux = -6.55e-11;
   // ...
+  fluxType = "KE+Pvector";
+  method = "VSS";
+  kineticFlux = -5.0e-6;
+  momentumFluxVector = (-2e-7, 0, 0);
 }
 ```
 
-The system responds to the thermal flux by developing a temperature
-gradient across the z-axis of the system. The thermal conductivity can
-be computed by relating the resulting thermal gradient to the imposed
-kinetic energy flux.
+along with definitions for *both* `kineticFlux` and `momentumFluxVector`.
 
-Since we have just two molecules in this example, openmd won't allow
-parallelization of more than 2 cores, so we can just run the following:
+Application of simultaneous momentum and kinetic energy fluxes result
+in both a velocity and thermal gradient response of the system,
+allowing for measurement of the shear viscosity of the fluid at a
+large number of temperature domains with one simulation.
+
+This simulation can be run with the following command:
 
[…]
