### 2024-10-09 #7004 @crdrisko
Hello @prgupta-cpfam, in an effort to head off any potential confusion with the [RNEMD sample](https://github.com/OpenMD/OpenMD/blob/main/samples/RNEMD/README.md), I've rewritten it with explicit commands and expected results for a number of new and existing simulations. In addition, some of the older, less interesting examples were removed, and results for longer simulations (1+ hrs.) were forgone. Hopefully this is what you were looking for, let me know if there's anything else you want to see from this sample.
