### 2024-10-11 6d10c56
Box volume issue in Current Density module
