### 2024-10-15 d18b2d7
Specifying minor version in build script
