+++ .github/workflows/build.yml
@@ -22,7 +22,7 @@ jobs:
       run: cmake --build . --target install
 
   ubuntu-builds:
-    runs-on: ubuntu-22
+    runs-on: ubuntu-22.04
 
     steps:
       - uses: actions/checkout@v4
