### 2024-10-16 f98d30c
Changing dependencies to match ubuntu 24 naming
