+++ .github/workflows/build.yml
@@ -22,14 +22,14 @@ jobs:
       run: cmake --build . --target install
 
   ubuntu-builds:
-    runs-on: ubuntu-22.04
+    runs-on: ubuntu-latest
 
     steps:
       - uses: actions/checkout@v4
 
       - name: Dependencies
         run: |
-          sudo apt-get install libfftw3-3 libfftw3-dev
+          sudo apt-get install libfftw3-bin libfftw3-dev
           sudo apt-get install openmpi-bin libopenmpi-dev
           sudo apt-get install python3-numpy python3-scipy
           sudo apt-get install qhull-bin libqhull-dev
