### 2024-10-18 #7004 @gezelter
@srmnitc Do you know what the published DOI for the article will be (when it is generated)?  We'd like to include a 'how to cite OpenMD' in the release version of the code.

Thanks!
