### 2024-10-18 ad34be2
Copyright headers updated for v3.1 release
