### 2024-10-18 #7004 @crdrisko
Hello @srmnitc, I can confirm that the authors, affiliations, and ORCIDs are up to date in the paper.  [Version 3.1](https://github.com/OpenMD/OpenMD/releases/tag/v3.1) is the new release with an affiliated Zenodo DOI: https://doi.org/10.5281/zenodo.13953130. We had to manually edit the Zenodo entry to reflect the information supplied in the paper (authors and license), but they all match now.
