+++ README.md
@@ -1,6 +1,8 @@
 # What is OpenMD?
 
-[![build](https://github.com/OpenMD/OpenMD/workflows/build/badge.svg)](https://github.com/OpenMD/OpenMD/actions?query=workflow%3Abuild) [![status](https://joss.theoj.org/papers/8841bf23a51ceaf3439f455219043855/status.svg)](https://joss.theoj.org/papers/8841bf23a51ceaf3439f455219043855)
+[![build](https://github.com/OpenMD/OpenMD/workflows/build/badge.svg)](https://github.com/OpenMD/OpenMD/actions?query=workflow%3Abuild) [![status](https://joss.theoj.org/papers/8841bf23a51ceaf3439f455219043855/status.svg)](https://joss.theoj.org/papers/8841bf23a51ceaf3439f455219043855) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.13953130.svg)](https://doi.org/10.5281/zenodo.13953130)
+
+
 
 
 OpenMD is an open source molecular dynamics engine which is capable of
