### 2024-10-21 #7004 @srmnitc
> @srmnitc Do you know what the published DOI for the article will be (when it is generated)? We'd like to include a 'how to cite OpenMD' in the release version of the code.
>
> Thanks!

No, unfortunately we wont have the DOI yet.
