### 2024-10-24 #7004 @srmnitc
@gezelter @crdrisko I am really sorry I did not respond before, I was unavailable. The guidelines mention that the Archive title should match that of the paper, since it is directly connected to the paper. Would this be okay with you? Once again, sorry for responding quicker.
