### 2024-10-24 OpenMD/OpenMD#29 @srmnitc
Changes to the paper
This is part of the review at https://github.com/openjournals/joss-reviews/issues/7004

Some small points for the paper:

- One of the authors is Charles F Vardeman II on the paper, but without II on the archive. Please confirm which is correct and update accordingly.
- line 15 and line 16: could be strengthened by a citation
- line 37: FAIR - how do you actually apply the FAIR principles to software? Or would you rather use the FAIR 4 Research Software publication (https://www.nature.com/articles/s41597-022-01710-x)
- line 53: could be strengthened with a citation
- line 73 and line 100: maybe a citation?

Please take a look, and feel free to change (or ignore) some of them as needed.
