### 2024-10-24 OpenMD/OpenMD#29 @gezelter
- Fixed archive title to match the paper.
- Archive now lists Chuck Vardeman's name as "Vardeman, Charles F., II" (His family name is Vardeman).
