+++ docs/paper/paper.bib
@@ -374,13 +374,21 @@
   year = 1998,
   url = {https://www.ccp5.ac.uk/newsletter-1998}
 }
-@article{Wilkinson2016,
-  author = {Wilkinson, Mark D and Dumontier, Michel and Aalbersberg, IJsbrand Jan and Appleton, Gabrielle and Axton, Myles and Baak, Arie and Blomberg, Niklas and Boiten, Jan-Willem and {da Silva Santos}, Luiz Bonino and Bourne, Philip E and Bouwman, Jildau and Brookes, Anthony J and Clark, Tim and Crosas, Merc{\`{e}} and Dillo, Ingrid and Dumon, Olivier and Edmunds, Scott and Evelo, Chris T and Finkers, Richard and Gonzalez-Beltran, Alejandra and Gray, Alasdair J G and Groth, Paul and Goble, Carole and Grethe, Jeffrey S and Heringa, Jaap and {'t Hoen}, Peter A C and Hooft, Rob and Kuhn, Tobias and Kok, Ruben and Kok, Joost and Lusher, Scott J and Martone, Maryann E and Mons, Albert and Packer, Abel L and Persson, Bengt and Rocca-Serra, Philippe and Roos, Marco and van Schaik, Rene and Sansone, Susanna-Assunta and Schultes, Erik and Sengstag, Thierry and Slater, Ted and Strawn, George and Swertz, Morris A and Thompson, Mark and van der Lei, Johan and van Mulligen, Erik and Velterop, Jan and Waagmeester, Andra and Wittenburg, Peter and Wolstencroft, Katherine and Zhao, Jun and Mons, Barend},
-  doi = {10.1038/sdata.2016.18},
-  journal = {Scientific Data},
-  number = {1},
-  pages = {160018},
-  title = {The {FAIR} Guiding Principles for scientific data management and stewardship},
-  volume = {3},
-  year = {2016}
+@article{Barker2022,
+author = {Barker, Michelle and {Chue Hong}, Neil P and Katz, Daniel S and Lamprecht, Anna-Lena and Martinez-Ortiz, Carlos and Psomopoulos, Fotis and Harrow, Jennifer and Castro, Leyla Jael and Gruenpeter, Morane and Martinez, Paula Andrea and Honeyman, Tom},
+doi = {10.1038/s41597-022-01710-x},
+journal = {Scientific Data},
+number = {1},
+pages = {622},
+title = {Introducing the {FAIR} Principles for research software},
+volume = {9},
+year = {2022}
+}
+@misc{Allen2017,
+author = {Allen, Michael P and Tildesley, Dominic J},
+doi = {10.1093/oso/9780198803195.001.0001},
+pages = {0},
+publisher = {Oxford University Press},
+title = {Computer Simulation of Liquids},
+year = {2017}
 }
