+++ .github/SECURITY.md
@@ -6,6 +6,7 @@ The following versions of OpenMD are currently being supported with security upd
 
 | Version | Supported          |
 | ------- | ------------------ |
+| 3.1.x   | :white_check_mark: |
 | 3.0.x   | :white_check_mark: |
 | 2.7.x   | :white_check_mark: |
 | < 2.7   | :x:                |
