### 2024-10-24 247b5fc
Missing reference and fixing bib formatting
