+++ docs/paper/paper.md
@@ -49,7 +49,7 @@ From the beginning, `OpenMD` has been an open source project, and has been maint
 
 # Statement of need
 
-`OpenMD` builds on the foundations of the Object-Oriented Parallel Simulation Engine (`OOPSE`) program [@Meineke2005], rewritten in modern C++. It differs from other contemporary molecular dynamics engines in its focus on complex interfaces. A number of features unique to `OpenMD` facilitate the study of this problem space, including efficient non-equilibrium algorithms, non-periodic simulations, metal polarizability models, and real space electrostatics. The first such feature is Reverse Non-Equilibrium Molecular Dynamics (RNEMD), a family of algorithms which impose a non-physical flux on a system and use linear response theory to compute transport properties as the system approaches steady state. The goal of RNEMD methods is to calculate the relevant transport property ($\lambda$) that connects the flux ($\textbf{J}$) and driving force ($\nabla X$) according to the generalized equation,
+`OpenMD` builds on the foundations of the Object-Oriented Parallel Simulation Engine (`OOPSE`) program [@Meineke2005], rewritten in modern C++. It differs from other contemporary molecular dynamics engines in its focus on complex interfaces. A number of features unique to `OpenMD` facilitate the study of this problem space, including efficient non-equilibrium algorithms, non-periodic simulations, metal polarizability models, and real space electrostatics. The first such feature is Reverse Non-Equilibrium Molecular Dynamics (RNEMD), a family of algorithms which impose a non-physical flux on a system and use linear response theory to compute transport properties as the system approaches steady state [@Zwanzig2001]. The goal of RNEMD methods is to calculate the relevant transport property ($\lambda$) that connects the flux ($\textbf{J}$) and driving force ($\nabla X$) according to the generalized equation,
 \begin{equation}\label{eq:linearConstitutive}
   \textbf{J} = - \lambda \nabla X.
 \end{equation}
@@ -58,7 +58,7 @@ From the beginning, `OpenMD` has been an open source project, and has been maint
 
 ## Reverse Non-Equilibrium Molecular Dynamics
 
-RNEMD methods impose a non-physical (heat, momentum, or particle) flux between different regions of the simulation. In response, the system develops a temperature, velocity, or concentration gradient between the two regions, and the linear coefficient connecting applied flux and measured gradient is a transport property of the material. Since the amount of the applied flux is known exactly in RNEMD, and the measurement of gradients is generally straightforward, imposed-flux methods typically take shorter simulation times to obtain converged results for transport properties, when compared with equilibrium MD or forward-NEMD approaches. If an interface lies between the two regions, these methods can also provide *interfacial* transport coefficients by mapping any spatial discontinuities in velocity, temperature, or concentration with the applied flux [@Kuang2012; @Stocker2014; @Drisko2024].
+RNEMD methods impose a non-physical (heat, momentum, or particle) flux between different regions of the simulation. In response, the system develops a temperature, velocity, or concentration gradient between the two regions, and the linear coefficient connecting applied flux and measured gradient is a transport property of the material. Since the amount of the applied flux is known exactly in RNEMD, and the measurement of gradients is generally straightforward, imposed-flux methods typically take shorter simulation times to obtain converged results for transport properties, when compared with equilibrium MD or forward-NEMD approaches. If an interface lies between the two regions, these methods can also provide *interfacial* transport coefficients by mapping any spatial discontinuities in concentration, velocity, or temperature with the applied flux [@Drisko2024; @Kuang2012; @Stocker2014].
 
[…]
