+++ docs/paper/paper.bib
@@ -375,20 +375,26 @@
   url = {https://www.ccp5.ac.uk/newsletter-1998}
 }
 @article{Barker2022,
-author = {Barker, Michelle and {Chue Hong}, Neil P and Katz, Daniel S and Lamprecht, Anna-Lena and Martinez-Ortiz, Carlos and Psomopoulos, Fotis and Harrow, Jennifer and Castro, Leyla Jael and Gruenpeter, Morane and Martinez, Paula Andrea and Honeyman, Tom},
-doi = {10.1038/s41597-022-01710-x},
-journal = {Scientific Data},
-number = {1},
-pages = {622},
-title = {Introducing the {FAIR} Principles for research software},
-volume = {9},
-year = {2022}
+  author = {Barker, Michelle and {Chue Hong}, Neil P and Katz, Daniel S and Lamprecht, Anna-Lena and Martinez-Ortiz, Carlos and Psomopoulos, Fotis and Harrow, Jennifer and Castro, Leyla Jael and Gruenpeter, Morane and Martinez, Paula Andrea and Honeyman, Tom},
+  doi = {10.1038/s41597-022-01710-x},
+  journal = {Scientific Data},
+  number = {1},
+  pages = {622},
+  title = {Introducing the {FAIR} Principles for research software},
+  volume = {9},
+  year = {2022}
 }
 @misc{Allen2017,
-author = {Allen, Michael P and Tildesley, Dominic J},
-doi = {10.1093/oso/9780198803195.001.0001},
-pages = {0},
-publisher = {Oxford University Press},
-title = {Computer Simulation of Liquids},
-year = {2017}
+  author = {Allen, Michael P and Tildesley, Dominic J},
+  doi = {10.1093/oso/9780198803195.001.0001},
+  publisher = {Oxford University Press},
+  title = {Computer Simulation of Liquids},
+  year = {2017}
+}
+@misc{Zwanzig2001,
+  author = {Zwanzig, Robert},
+  doi = {10.1093/oso/9780195140187.001.0001},
+  publisher = {Oxford University Press},
+  title = {Nonequilibrium Statistical Mechanics},
+  year = {2001}
 }
