+++ docs/paper/paper.md
@@ -45,11 +45,11 @@ Molecular dynamics (MD) simulations help bridge the gap between quantum mechanic
 
 `OpenMD` is capable of efficiently simulating a variety of complex systems using standard point-particle atoms, as well as atoms with orientational degrees of freedom (e.g. point multipoles and coarse-grained assemblies), and atoms with additional degrees of freedom (e.g. fluctuating charges). It provides a test-bed for new molecular simulation methodology, while being efficient and easy to use. Liquids, proteins, zeolites, lipids, inorganic nanomaterials, transition metals (bulk, flat interfaces, and nanoparticles), alloys, solid-liquid interfaces, and a wide array of other systems have all been simulated using this code. `OpenMD` works on parallel computers using the Message Passing Interface (MPI), and is packaged with trajectory analysis and utility programs that are easy to use, extend, and modify.
 
-From the beginning, `OpenMD` has been an open source project, and has been maintained in accordance with the FAIR (findable, accessible, interoperable, and reusable) principles for research software [@Barker2022]. It uses a robust meta-data language that is tightly integrated with input (`.omd`) and trajectory (`.dump`) files, providing a standardized, human-readable way to completely describe molecular systems and simulation parameters. This allows `OpenMD` simulations to be easily reproduced, modified, and reused by others. The `<MetaData>` section of these files also serve as a form of rich, machine-actionable meta-data, clearly specifying the composition of the molecular system in a manner that promotes interoperability with other software tools. Additionally, all data files are stamped with the code revision that generated that data, greatly aiding reproducibility.
+From the beginning, `OpenMD` has been an open source project and has been maintained in accordance with the FAIR (findable, accessible, interoperable, and reusable) principles for research software [@Barker2022]. It uses a meta-data language that is tightly integrated with input (`.omd`) and trajectory (`.dump`) files, providing a standardized, human-readable way to completely describe molecular systems and simulation parameters. All data files are stamped with the code revision that generated that data. This allows `OpenMD` simulations to be easily reproduced, modified, and reused by others. The `<MetaData>` section of these files also serve as a form of machine-actionable meta-data, clearly specifying the composition of the molecular system in a manner that promotes interoperability with other software tools.
 
 # Statement of need
 
-`OpenMD` builds on the foundations of the Object-Oriented Parallel Simulation Engine (`OOPSE`) program [@Meineke2005], rewritten in modern C++. It differs from other contemporary molecular dynamics engines in its focus on complex interfaces. A number of features unique to `OpenMD` facilitate the study of this problem space, including efficient non-equilibrium algorithms, non-periodic simulations, metal polarizability models, and real space electrostatics. The first such feature is Reverse Non-Equilibrium Molecular Dynamics (RNEMD), a family of algorithms which impose a non-physical flux on a system and use linear response theory to compute transport properties as the system approaches steady state [@Zwanzig2001]. The goal of RNEMD methods is to calculate the relevant transport property ($\lambda$) that connects the flux ($\textbf{J}$) and driving force ($\nabla X$) according to the generalized equation,
[…]
