+++ docs/paper/paper.bib
@@ -391,10 +391,3 @@
   title = {Computer Simulation of Liquids},
   year = {2017}
 }
-@misc{Zwanzig2001,
-  author = {Zwanzig, Robert},
-  doi = {10.1093/oso/9780195140187.001.0001},
-  publisher = {Oxford University Press},
-  title = {Nonequilibrium Statistical Mechanics},
-  year = {2001}
-}
