+++ .github/SECURITY.md
@@ -8,8 +8,7 @@ The following versions of OpenMD are currently being supported with security upd
 | ------- | ------------------ |
 | 3.1.x   | :white_check_mark: |
 | 3.0.x   | :white_check_mark: |
-| 2.7.x   | :white_check_mark: |
-| < 2.7   | :x:                |
+| < 3.0   | :x:                |
 
 ## Reporting a Vulnerability
 
