### 2024-10-25 #7004 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
