### 2024-10-31 #7004 @Kevin-Mattheus-Moerman
@gezelter as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I will now process some final checks:

- [ ] Please add the country name (in full, do not use country acronyms) to each affiliation.
