+++ docs/paper/paper.md
@@ -25,15 +25,15 @@ authors:
     corresponding: true
     affiliation: 1
 affiliations:
- - name: Department of Chemistry and Biochemistry, University of Notre Dame, Notre Dame, IN, United States of America
+ - name: Department of Chemistry and Biochemistry, University of Notre Dame, Notre Dame, IN, United States
    index: 1
- - name: Department of Physics, Goshen College, Goshen, IN, United States of America
+ - name: Department of Physics, Goshen College, Goshen, IN, United States
    index: 2      
- - name: Department of Chemistry, Oklahoma State University, Stillwater, OK, United States of America
+ - name: Department of Chemistry, Oklahoma State University, Stillwater, OK, United States
    index: 3
- - name: Department of Biochemistry, Chemistry, Environment, and Physics, Suffolk University, Boston, MA, United States of America
+ - name: Department of Biochemistry, Chemistry, Environment, and Physics, Suffolk University, Boston, MA, United States
    index: 4
- - name: Center for Research Computing, University of Notre Dame, Notre Dame, IN, United States of America
+ - name: Center for Research Computing, University of Notre Dame, Notre Dame, IN, United States
    index: 5   
 date: 25 June 2024
 bibliography: paper.bib
@@ -43,7 +43,9 @@ bibliography: paper.bib
 
 Molecular dynamics (MD) simulations help bridge the gap between quantum mechanical calculations, which trade computational resources and system size for chemical accuracy, and continuum simulations, which utilize bulk and surface material properties to make predictions. MD provides the ability to study emergent properties and dynamics for relatively large systems ($\sim 10^6$ atoms) over a span of nano- to micro-seconds. MD has been used for simulations of complex systems such as liquids, proteins, membranes, nanoparticles, interfaces, and porous materials, and the approximations used in MD simulations allow us to reach experimentally-relevant time and length scales [@Allen2017]. A molecular dynamics engine leverages classical equations of motion to evolve atomic or molecular coordinates according to a well-defined potential energy function, known as a force field, which is a function of those atomic coordinates. There are a number of high quality molecular dynamics engines that specialize in materials [@LAMMPS] or biomolecular [@CHARMM; @AmberTools] simulations or are models of computational efficiency [@GROMACS; @OpenMM]. In this paper, we provide background on an open source molecular dynamics engine, `OpenMD`, which specializes in complex systems and interfaces and was just released into version 3.1.
 
[…]
