### 2024-11-01 6dc8738
Create CITATION.cff
