### 2024-11-15 #7004 @Kevin-Mattheus-Moerman
@openjournals/dev @arfon @xuanxu how do you think we should handle this? The reviewer link listed on the JOSS website for this paper now goes to an error 404. Should we update this reviewers handle here, and then re-accept, so the reviewer handle listed on the website will be accurate? If so, what is the best way of doing that. Should I use the editorialbot to remove the old one and add the new one, and create a completed checklist for the new one (or manually edit the current)? Thanks
